% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas.balance.R
\name{cas.balance}
\alias{cas.balance}
\title{Do mass balance calculation}
\usage{
cas.balance(datain, element, phase, mode, density)
}
\arguments{
\item{datain}{Element abundances with row "WR"}

\item{element}{Vector of string such for c("Ba", "Nb", "La"...)}

\item{phase}{Vector of name of phases such for c("ol", "pig")}

\item{mode}{Vector of modal abundance of phases with label}

\item{density}{Vector of density of phases with label}
}
\value{
Element abundances with rows of remainder and series of
    quantity, and columns of density and mode
}
\description{
Do mass balance calculation
}

