% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas.normalize.R
\name{cas.normalize}
\alias{cas.normalize}
\title{Return normalized element abundances.  Note that only elements that
exist both in castbl and ref are processed.  See also
"Geochemical Modelling..." by Janousek et al. (2015)}
\usage{
cas.normalize(castbl, ref, suffix_after_name_of_element = NULL)
}
\arguments{
\item{castbl}{A dataframe of element abundances}

\item{ref}{A numeric vector of element abundances with label}
}
\value{
A ref-normalized daraframe with only elements defined in
    ref
}
\description{
Return normalized element abundances.  Note that only elements that
exist both in castbl and ref are processed.  See also
"Geochemical Modelling..." by Janousek et al. (2015)
}

