% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas.path.R
\name{cas.path}
\alias{cas.path}
\title{Convert filename of local library to absolute path}
\usage{
cas.path(name)
}
\arguments{
\item{name}{A filename}
}
\value{
A path to lib directory
}
\description{
Convert filename of local library to absolute path
}
\examples{
cas.path("periodic-table.csv")
}

