% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas.ref.R
\name{cas.ref}
\alias{cas.ref}
\title{Return element abundances from file `ref.csv'}
\usage{
cas.ref(analysis, property = NULL)
}
\arguments{
\item{analysis}{A name of REFERENCE that is either "Boynton 1989", "McDonough 1995", or "Wasson 1988"}

\item{property}{A numeric vector of element property with label
that serves as sort key (default=NULL).  You can feed output from `cas.periodic'}
}
\value{
A numeric vector of element abundances with label
}
\description{
Return element abundances from file `ref.csv'
}
\examples{
cas.ref("Boynton 1989")
cas.ref("Wasson 1988",cas.periodic('atomicnumber'))
cas.ref("McDonough 1995",cas.periodic('volatility'))
}

