% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castbl.path.R
\name{castbl.path}
\alias{castbl.path}
\title{Convert filename of local library to absolute path}
\usage{
castbl.path(name)
}
\arguments{
\item{name}{A filename}
}
\value{
A path to lib directory
}
\description{
Convert filename of local library to absolute path
}
\examples{
castbl.path("periodic-table.csv")
}

