% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.casteml.download.R
\name{cbk.casteml.download}
\alias{cbk.casteml.download}
\title{Download analysis record to CASTEML file}
\usage{
cbk.casteml.download(arguments)
}
\arguments{
\item{arguments}{Arguments that will pass to `casteml download'.
Typically unique indentification number of stones in Medusa.}
}
\value{
file path to CASTEML file that was downloaded in temporary
  directory
}
\description{
Download analysis record to CASTEML file.  The
  downloaded is stored in a temporary directory.  This function
  returns path to the file.
}
\examples{
stone <- c("20080616170000.hk","20080616170056.hk","20080616170054.hk")
pmlfile <- cbk.casteml.download(stone)

pmlfile <- cbk.casteml.download("20081202172326.hkitagawa")
}
\seealso{
\code{casteml download}, \url{https://github.com/misasa/casteml}, \code{\link{cbk.casteml.convert}}
}

