% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.category.dwim.R
\name{cbk.category.dwim}
\alias{cbk.category.dwim}
\title{Return representative category}
\usage{
cbk.category.dwim(pmlame)
}
\arguments{
\item{pmlame}{A dataframe with rows of stone and columns of chem}
}
\value{
A representative category in the dataframe
}
\description{
Return representative category
}
\examples{
pmlfile <- cbk.download.casteml("20081202172326.hkitagawa")
pmlame  <- cbk.read.casteml(pmlfile,"ppm",category=NULL)
category <- cbk.category.dwim(pmlame)
}

