% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.category.R
\name{cbk.category}
\alias{cbk.category}
\title{Return list of categories that are acceptable by `cbk.plot'}
\usage{
cbk.category()
}
\value{
A list of categories that are acceptable by `cbk.plot'
}
\description{
Return list of categories that are acceptable by `cbk.plot'
}
\examples{
categorylist <- cbk.category()
}

