% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.category.suggest.R
\name{cbk.category.suggest}
\alias{cbk.category.suggest}
\title{Suggest a category to plot}
\usage{
cbk.category.suggest(pmlame)
}
\arguments{
\item{pmlame}{A dataframe with rows of stone and columns of chem}
}
\value{
A representative category in the dataframe
}
\description{
Suggest a category to plot
}
\examples{
pmlfile <- cbk.path("20081202172326.hkitagawa.pml")
pmlame  <- cbk.read.casteml(pmlfile,"ppm",category=NULL)
category <- cbk.category.suggest(pmlame)
}

