% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.convector.R
\name{cbk.convector}
\alias{cbk.convector}
\title{Return conversion cofficient for unit}
\usage{
cbk.convector(tableunit)
}
\arguments{
\item{tableunit}{Unit string, that is one of
none, g/g, wt\%, cg/g, \%, permil, mg/g, ppm, ug/g, ppb, ng/g, pg/g}
}
\value{
Conversion cofficient.  When "permil", this returns 1000
  instead of 0.001
}
\description{
Return conversion cofficient for unit
}
\examples{
cbk.convector("permil")
}

