% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.convert.casteml.R
\name{cbk.convert.casteml}
\alias{cbk.convert.casteml}
\title{Convert CASTEML file to CSV file}
\usage{
cbk.convert.casteml(pmlfile, ..., category = NULL)
}
\arguments{
\item{pmlfile}{file path to CASTEML file}

\item{category}{category to pass to `casteml convert'}
}
\value{
file path to converted file
}
\description{
Convert CASTEML file to CSV file.  This function
  returns path to the converted file.  The converted file is stored
  in a temporary directory.
}
\examples{
pmlfile <- cbk.download.casteml("20081202172326.hkitagawa")
cbkfile <- cbk.convert.casteml(pmlfile,category="trace")

cbkfile <- cbk.convert.casteml(cbk.path("20081202172326.hkitagawa.pml"),category="trace")
}
\seealso{
\code{casteml convert}, \url{https://github.com/misasa/casteml}, \code{\link{cbk.download.casteml}}
}

