% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.disp.dummy.R
\name{cbk.disp.dummy}
\alias{cbk.disp.dummy}
\title{Display dummy diagram}
\usage{
cbk.disp.dummy(error, xlabel = NULL)
}
\arguments{
\item{error}{A error message to be displayed on diagram}
}
\value{
A dataframe used to plot
}
\description{
Display dummy diagram
}
\examples{
tryCatch({plot(XX,YY)
},error=function(error){
cbk.disp.dummy(error)})
}

