% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.download.casteml.1.R
\name{cbk.download.casteml.1}
\alias{cbk.download.casteml.1}
\title{Download analysis records as a CASTEML file via HTTP}
\usage{
cbk.download.casteml.1(stone, file = NULL, force = FALSE,
  directAuth = NULL, Recursive = FALSE, recursive = FALSE)
}
\arguments{
\item{stone}{Unique identification number of a stone in Medusa.}

\item{file}{File path to save downloaded CASTEML file.  Default is
a temporary file in a temporary directory.}

\item{force}{Force download CASTEML file.}

\item{directAuth}{Parameter for HTTP basic authentication.}

\item{Recursive}{Download a whole family of a stone (equivalent to
--family in casteml download)}

\item{recursive}{Download descendants of a stone (equivalent to
--descendant in casteml download).}
}
\value{
Path to CASTEML file that was downloaded.
}
\description{
Download analysis records as a CASTEML file via HTTP.
  This function returns path to the file.  The file is stored in a
  temporary directory unless specified.  Note with the same
  arguments, this function downloads only once per an R session.
}
\details{
This function was created on 2017-04-11 to download
  CASTEML file by its own from Medusa and the first argument can
  only accept single `stone' unlike \link{cbk.download.casteml} can
  accept an array consists of stones.
}
\examples{
stone <- c("20080616170000.hk","20080616170056.hk","20080616170054.hk")
pmlfiles <- lapply(stone, cbk.download.casteml.1)

pmlfile <- cbk.download.casteml.1("20081202172326.hkitagawa")
directAuth <- list(uri="dream.misasa.okayama-u.ac.jp/demo/",user='admin',password='admin')
pmlfile <- cbk.download.casteml.1("20110416134901-075-241",directAuth=directAuth)
pmlfile <- cbk.download.casteml.1("20110416134901-075-241",directAuth=directAuth,recursive=TRUE)
pmlfile <- cbk.download.casteml.1("20110416134901-075-241",directAuth=directAuth,Recursive=TRUE)
}
\seealso{
\code{casteml download},
  \url{https://github.com/misasa/casteml}, and
  \code{\link{cbk.convert.casteml}}
  \code{\link{cbk.download.casteml}}
}
