% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.download.pmlame.R
\name{cbk.download.pmlame}
\alias{cbk.download.pmlame}
\title{Download analysis records from Medusa and return a pmlame}
\usage{
cbk.download.pmlame(stone, opts = NULL, file = NULL, tableunit = "none",
  force = FALSE, verbose = TRUE)
}
\arguments{
\item{stone}{Unique identification number of stones in Medusa.}

\item{file}{File path to save downloaded TBLAME.csv.  Default is
a temporary file in a temporary directory.}

\item{tableunit}{Output unit that will be resolved by
cbk.convector() (default: "none").}

\item{force}{Force download a pmlame (default: FALSE).}
}
\value{
A pmlame with unit organized.
}
\description{
Download analysis records from Medusa and return a
  pmlame.  The pmlame is stored in a temporary directory as a
  TBLAME.csv unless specified.  Note with the same arguments, this
  function downloads only once per an R session.
}
\examples{
pmlame <- cbk.download.pmlame("20081202172326.hkitagawa")
}
\seealso{
\code{\link{cbk.download.casteml}}, and
  \url{https://github.com/misasa/casteml}
}
