% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.exec.R
\name{cbk.exec}
\alias{cbk.exec}
\title{Open file even on Mac.  This is copy from
http://stackoverflow.com/questions/12273346/system-independent-method-of-opening-afile}
\usage{
cbk.exec(x)
}
\arguments{
\item{x}{A filename}
}
\value{
result
}
\description{
Open file even on Mac.  This is copy from
http://stackoverflow.com/questions/12273346/system-independent-method-of-opening-afile
}

