% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.filter.colmeans.R
\name{cbk.filter.colmeans}
\alias{cbk.filter.colmeans}
\title{Return mean value of each element}
\usage{
cbk.filter.colmeans(pmlame)
}
\arguments{
\item{pmlame}{A dataframe with rows of stone and columns of chem}
}
\value{
A dataframe with mean value of each column
}
\description{
Return mean value of each element
}
\examples{
pmlame  <- cbk.read.casteml(cbk.path("20081202172326.hkitagawa.pml"))
pmlame0 <- pmlame[,colnames(pmlame) != "sample_id"]
pmlame1 <- cbk.filter.colmeans(pmlame0)
}

