% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.filter.drop.dharma.R
\name{cbk.filter.drop.dharma}
\alias{cbk.filter.drop.dharma}
\title{Remove rows where all values are NA}
\usage{
cbk.filter.drop.dharma(pmlame, column = FALSE)
}
\arguments{
\item{pmlame}{A dataframe with rows of stone and columns of chem}

\item{column}{Flag to remove columns}
}
\value{
A dataframe without removed rows
}
\description{
Remove rows where all values are NA
}
\examples{
pmlfile <- cbk.download.casteml("20081202172326.hkitagawa")
pmlame  <- cbk.read.casteml(pmlfile,"ppm",category=NULL)
pmlame1 <- cbk.filter.drop.dharma(pmlame)
}

