% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.filter.reduce.R
\name{cbk.filter.reduce}
\alias{cbk.filter.reduce}
\title{Convert major-element concentration as oxide to one as metal}
\usage{
cbk.filter.reduce(pmlame)
}
\arguments{
\item{pmlame}{A pmlame of element abundances (with columns of stone and rows of chem)}
}
\value{
A pmlame with converted columns
}
\description{
Convert major-element concentration as oxide to one as metal
}
\examples{
pmlfile <- cbk.path("20081202172326.hkitagawa.pml")
pmlame  <- cbk.read.casteml(pmlfile,"ppm",category=NULL)
pmlame1 <- cbk.filter.reduce(pmlame)
}

