% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.iso.R
\name{cbk.iso}
\alias{cbk.iso}
\title{Return isotope information}
\usage{
cbk.iso()
}
\value{
A pmlame of isotope information
}
\description{
Return isotope information.  It is in a form of pmlame.
}
\examples{
cbk.iso()['Li7','ratio']
cbk.iso()['Li7','weight']
}
\seealso{
\code{\link{cbk.periodic}}
}

