% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.atomify.R
\name{cbk.lame.atomify}
\alias{cbk.lame.atomify}
\title{Estimate element abundances from ion-intensity obtained by
  mass spectrometry.}
\usage{
cbk.lame.atomify(pmlame, ionic_ratio, ion_yield)
}
\arguments{
\item{pmlame}{A pmlame with internal-referene element such for Si
and Ca.  Note that you have to reduce SiO2 in advance.}

\item{ionic_ratio}{A pmlame-like ion signal intensities relative
to internal-reference isotope.  For example, I(Li7)/I(Si29).}

\item{ion_yield}{Sensitivities of element that were determined by
analyses of several reference materials.}
}
\value{
A pmlame of element abundances.
}
\description{
Estimate element abundances from ion-intensity
  obtained by mass spectrometry.  Inputs are element abundance of
  internal-reference element, ionic_ratio, and ion_yield.
}
\details{
\itemize{
  \item{ion_yield    = ionic_ratio / atomic_ratio}
  \item{ionic_ratio  = I(Li7+)      / I(Si29+)}
  \item{atomic_ratio = [Li7]        / [Si29]}
  \item{[Li7]        = [Li]/m!(Li7)}
  \item{[Si29]       = [Si]/m!(Si29)}
  \item{m!(Li7)      = m(Li)/R(Li7)}
  \item{m!(Si9)      = m(Si)/R(Si29)}
  \item{[Li]         <- ionic_ratio/ion_yield * [Si]/m!(Si29) * m!(Li7)}
}
}

