% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.colMeans.R
\name{cbk.lame.colMeans}
\alias{cbk.lame.colMeans}
\title{Return mean of each chem}
\usage{
cbk.lame.colMeans(pmlame, stonelist_regexp = NULL)
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{stonelist_regexp}{A regexp string to split stone name}
}
\value{
A pmlame with mean value of each column
}
\description{
Return mean of each chem
}
\examples{
pmlame  <- cbk.read.casteml(cbk.path("20130528105235-594267.pml"))
pmlame0 <- pmlame[,setdiff(colnames(pmlame), c("sample_id","image_id"))]
pmlame1 <- cbk.lame.colMeans(pmlame0)
}
