% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.colSds.R
\name{cbk.lame.colSds}
\alias{cbk.lame.colSds}
\title{Return standard deviation of each chem}
\usage{
cbk.lame.colSds(pmlame, stonify_regexp = "@[[:digit:]]+$")
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{stonify_regexp}{A regexp string to be stripped off to stonefy
analyses. Matched string will be stripped off from acqlist to
have stonelist.}
}
\value{
A pmlame with standard deviation value of each column
}
\description{
Return standard deviation of each chem
}
\examples{
pmlame  <- cbk.read.casteml(cbk.path("20130528105235-594267.pml"))
pmlame0 <- cbk.lame.regulate(pmlame,mean=TRUE,error=FALSE,extra=FALSE)
pmlame1 <- cbk.lame.colSds(pmlame0)
}
