% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.delivery.R
\name{cbk.lame.delivery}
\alias{cbk.lame.delivery}
\title{Estimate delivery rate of element from sputter rate}
\usage{
cbk.lame.delivery(pmlame, intlame, sputter.rate = 1.1e-09, verbose = FALSE)
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{intlame}{A pseudo-pmlame of ion intensity with row of stone
and column of isomeas [cps].  Typically rowname is with
`@acq-number'.}

\item{sputter.rate}{Sputter rate [g/s] (default: 1.1e-9 that
corresponds to abbration of 40 s to produce a pit with diameter
of 25 micron, depth of 30 micron, and density of 3 g/cc).}

\item{verbose}{Output debug info (default: FALSE).}
}
\value{
Delivery rate of element from solid target to detector.
}
\description{
Estimate delivery rate of element from sputter rate.
  Calculation is made in dimension of CGS.
}
\details{
With ion Si29 intensity 3.85e+05 [cps], element SiO2
  abundance 59 wt%, and sputter rate 1.1 ng/s, the delivery rate is
  estimated to be 1.25 ppm.
}
\examples{
pmlame0      <- data.frame(row.names=c("ref-gl-tahiti","ref-cpx-klb1"),SiO2=c(0.590,0.520),Li=c(20.8e-06,1.37e-06),Sr=c(107e-06,36.1e-06))
pmlame1      <- cbk.lame.reduce(pmlame0)
intlame0     <- data.frame(row.names=c("ref_gl_tahiti@1","ref_cpx_klb1@1"),Si29=c(3.85e+05,3.95e+05),Li7=c(1.87e+03,1.73e+02),Sr88=c(3.24e+04,1.50e+04))
radius       <- 12.5e-6 # Radius of a pit [m]
depth        <- 30e-6   # Depth of a pit [m]
t            <- 40      # Period of irradiation [s]
density      <- 3       # Density of solid target [g/cc]
sputter.rate <- 3.14 * (radius * 100)^2 * (depth * 100) * density / t # [g/s]
cbk.lame.delivery(pmlame1,intlame0,sputter.rate)
}
