% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.drop.stone.R
\name{cbk.lame.drop.stone}
\alias{cbk.lame.drop.stone}
\title{Drop stone from pmlame}
\usage{
cbk.lame.drop.stone(pmlame, stone, verbose = FALSE)
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{stone}{List of stone to drop from pmlame}

\item{verbose}{Output debug info (default: FALSE).}
}
\value{
A pmlame without the specified stone
}
\description{
Drop stone from pmlame
}
\examples{
pmlame <- data.frame(row.names=c("stone@1","stone@2","stone@3"), SiO2=c(0.525,0.420,0.332), fo=c(0.353,3.400,20.323), d7Li=c(1.0,3.3,-1.0))
cbk.lame.drop.stone(pmlame,"stone@1")
cbk.lame.drop.stone(pmlame,c("stone@1","stone@2"))
}
\seealso{
\code{cbk.lame.drop.stone}
}
