% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.merge.error.R
\name{cbk.lame.merge.error}
\alias{cbk.lame.merge.error}
\title{Merge meanlame and errorlame to pmlame interleavely}
\usage{
cbk.lame.merge.error(meanlame, errorlame, verbose = FALSE)
}
\arguments{
\item{meanlame}{A pmlame of mean values with colname of chem ('Li7').}

\item{errorlame}{An errorlame, that is pmlame of error values with colname of chem ('Li7' instead of 'Li7_error').}

\item{verbose}{Output debug info (default: FALSE).}
}
\value{
A proper pmlame with both mean and error columns.
}
\description{
Merge meanlame and errorlame to pmlame interleavely
}
\examples{
pmlame     <- cbk.read.casteml(cbk.path("20130528105235-594267.pml"))
pmlame0    <- pmlame[,!grepl("_error",colnames(pmlame))]
pmlame1    <- pmlame0[,setdiff(colnames(pmlame0), c("sample_id","image_id"))]
meanlame1  <- cbk.lame.colMeans(pmlame1)
errorlame1 <- cbk.lame.colSds(pmlame1)
pmlame2    <- cbk.lame.merge.error(meanlame1,errorlame1)
}
\seealso{
\code{cbk.lame.error}
}
