% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.normalize.R
\name{cbk.lame.normalize}
\alias{cbk.lame.normalize}
\title{Return normalized element abundances}
\usage{
cbk.lame.normalize(pmlame, reflame, suffix_after_chem = NULL,
  verbose = FALSE)
}
\arguments{
\item{pmlame}{A pmlame of element abundances of sample.}

\item{reflame}{A pmlame of element abundances of reference.}

\item{suffix_after_chem}{String to recognize column of errors.
Feed "_error" when necessary (default: NULL).}

\item{verbose}{Output debug info (default: FALSE).}
}
\value{
A ref-normalized daraframe with only elements defined in
  ref.
}
\description{
Return normalized element abundances.  Note that only
  elements that exist both in pmlame and reflame are processed.
  See also "Geochemical Modelling..." by Janousek et al. (2015)
}
\examples{
pmlame  <- cbk.read.dflame(cbk.path("20081202172326.hkitagawa_trace.dflame"),"ppm")
reflame <- cbk.ref("Boynton.1989","ppm")
cbk.lame.normalize(pmlame,reflame)
}
\seealso{
\code{\link{cbk.ref}} and \code{\link{cbk.periodic}}
}
