% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.regulate.R
\name{cbk.lame.regulate}
\alias{cbk.lame.regulate}
\title{Regulate pmlame to remove non-chem components}
\usage{
cbk.lame.regulate(pmlame, chem = TRUE, error = TRUE, extra = FALSE)
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{chem}{Flag to extract chem columns (default: TRUE)}

\item{error}{Flag to extract error columns (default: TRUE)}

\item{extra}{Flag to extract extra columns (default: FALSE)}
}
\value{
A pmlame with only chem
}
\description{
Regulate pmlame to remove non-chem components
}
\examples{
pmlame0 <- cbk.read.casteml(cbk.path("20130528105235-594267.pml"))
pmlame  <- cbk.lame.regulate(pmlame0)
}

