% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.rep.R
\name{cbk.lame.rep}
\alias{cbk.lame.rep}
\title{Replicate pmlame in horizontal or vertical direction}
\usage{
cbk.lame.rep(pmlame, times, direction = NULL)
}
\arguments{
\item{pmlame}{A pmlame to be replicated}

\item{times}{Number of replication}

\item{direction}{direction to make replicate (`c' or `r').  This
function guesses the direction to be replicated when pmlame is
with single row or column.}
}
\value{
A pmlame with repeated column or row.
}
\description{
Replicate pmlame in horizontal or vertical direction
}
\examples{
pmlame0 <- data.frame(row.names=c("stone.1","stone.2"), SiO2=c(0.525,0.420))
cbk.lame.rep(pmlame0,3)
}

