% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.show.R
\name{cbk.lame.show}
\alias{cbk.lame.show}
\title{Show pmlame in console}
\usage{
cbk.lame.show(pmlame, format = NULL)
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{format}{Format specifier that is passed to \code{casteml
convert}.}
}
\value{
A pmlame that is not processed.
}
\description{
Show pmlame in console
}
\examples{
pmlame0 <- structure(list(SiO2 = c(0.59, 0.52), Li = c(2.08e-05, 1.37e-06), Sr = c(0.000107, 3.61e-05)), row.names = c("ref-gl-tahiti", "ref-cpx-klb1"), .Names = c("SiO2", "Li", "Sr"), class = "data.frame")
cbk.lame.show(pmlame0)
cbk.lame.show(pmlame0,'isorg')
}

