% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.texify1.R
\name{cbk.lame.texify1}
\alias{cbk.lame.texify1}
\title{Convert pmlame to texfile}
\usage{
cbk.lame.texify1(pmlame, chem, outfile = NULL, verbose = FALSE)
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{chem}{List of chem such as c("Li","Si","Ca","Ca.1","Rb").}

\item{outfile}{File path to texfile.}

\item{verbose}{Output debug info.}
}
\value{
File path to texfile.
}
\description{
Convert pmlame to texfile.
}
\examples{
pmlame <- cbk.read.casteml(cbk.path("ref_phase_obj.pml"))
chem <- c("Li","B","Si","Rb","Sr","Y","Zr","Nb","Cs")
cbk.lame.texify1(pmlame,chem)
}
