% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.normalize.R
\name{cbk.normalize}
\alias{cbk.normalize}
\title{Return normalized element abundances}
\usage{
cbk.normalize(cbktbl, ref, suffix_after_name_of_element = NULL)
}
\arguments{
\item{cbktbl}{A dataframe of element abundances}

\item{ref}{A numeric vector of element abundances with label}

\item{suffix_after_name_of_element}{String to recognize column of
errors.  Feed "_error" when necessary (default: NULL)}
}
\value{
A ref-normalized daraframe with only elements defined in
  ref
}
\description{
Return normalized element abundances.  Note that only
  elements that exist both in cbktbl and ref are processed.  See
  also "Geochemical Modelling..." by Janousek et al. (2015)
}
\examples{
cbk.normalize(cbk.read.dataframe(cbk.path("20081202172326.kitagawa_trace.dataframe"),"ppm"),cbk.ref("Boynton.1989"))
}
\seealso{
\code{\link{cbk.ref}}, \code{\link{cbk.periodic}}
}

