% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.path.R
\name{cbk.path}
\alias{cbk.path}
\title{Return absolute path of files stored in local library}
\usage{
cbk.path(filename)
}
\arguments{
\item{filename}{A filename}
}
\value{
An absolute path of files stored in local library
}
\description{
Return absolute path of files stored in local library
}
\examples{
cbkfile <- cbk.path("periodic-table1.dataframe")
cbkfile <- cbk.path("ref1.dataframe")
pmlfile <- cbk.path("20081202172326.kitagawa.pml")
cbkfile <- cbk.path("20081202172326.kitagawa_trace.dataframe")
}

