% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.periodic.R
\name{cbk.periodic}
\alias{cbk.periodic}
\title{Return properties of elements from a file `periotic-table.csv'}
\usage{
cbk.periodic(property)
}
\arguments{
\item{property}{A name of PROPERTY such for atomicnumber,
volatility, and compatibility}
}
\value{
A numeric vector of element property with label
}
\description{
Return properties of elements from a file `periotic-table.csv'
}
\examples{
cbk.periodic("atomicnumber")
}

