% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.lead.R
\name{cbk.plot.lead}
\alias{cbk.plot.lead}
\title{Read CASTEML file and create lead diagram}
\usage{
cbk.plot.lead(pmlame)
}
\arguments{
\item{pmlame}{A dataframe of element abundances (or pmlfile or stone-ID)}
}
\value{
Dataframe used to plot the diagram
}
\description{
Read CASTEML dataframe and create lead diagram.  This
  function does not save the created diagram.  You should prepare a
  canvas in advance.
}
\examples{
pmlfile <- cbk.path("20081202172326.hkitagawa.pml")
pmlame  <- cbk.read.casteml(pmlfile,category="lead")
cbk.plot.lead(pmlame)
}
\seealso{
\url{https://github.com/misasa/casteml}
}

