% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.lithium.R
\name{cbk.plot.lithium}
\alias{cbk.plot.lithium}
\title{Read CASTEML file and create lithium diagram}
\usage{
cbk.plot.lithium(pmlfile, tableunit = "none")
}
\arguments{
\item{pmlfile}{File path to CASTEML file}

\item{tableunit}{Unit to toss to cbk.read.casteml().  This function
reads both element concentration and isotope composition, thus,
any unit cannot be appropriate as of November 15 (2016).  Do not
feed `ppm.'}
}
\value{
Dataframe used to plot the diagram
}
\description{
Read CASTEML file and create lithium diagram.  This
  function does not save the created diagram.  You should prepare a
  canvas in advance.
}
\examples{
pmlfile <- cbk.download.casteml(c("-r","20130528105235-594267"))
cbk.plot.lithium(pmlfile)
}
\seealso{
\url{https://github.com/misasa/casteml}
}

