% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.message.R
\name{cbk.plot.message}
\alias{cbk.plot.message}
\title{Display dummy diagram}
\usage{
cbk.plot.message(pmlame, msg)
}
\arguments{
\item{msg}{A error message to be displayed on diagram}
}
\value{
A dataframe used to plot
}
\description{
Display dummy diagram
}
\examples{
pmlame <- cbk.read.casteml(cbk.path("20081202172326.hkitagawa.pml"),"ppm")
tryCatch({plot(XX,YY)
},error=function(e){
cbk.plot.message(pmlame,e)})
}

