% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.oxygen.R
\name{cbk.plot.oxygen}
\alias{cbk.plot.oxygen}
\title{Read CASTEML file and create oxygen diagram}
\usage{
cbk.plot.oxygen(pmlame, tableunit = "none")
}
\arguments{
\item{pmlame}{A dataframe of element abundances}

\item{tableunit}{Unit to toss to cbk.read.casteml()}
}
\value{
Dataframe used to plot the diagram
}
\description{
Read CASTEML file and create oxygen diagram.  This
  function does not save the created diagram.  You should prepare a
  canvas in advance.
}
\examples{
pmlfile <- cbk.download.casteml(c("-r","20130528105235-594267"))
pmlame  <- cbk.read.casteml(pmlfile,category="oxygen")
cbk.plot.oxygen(pmlame)
}
\seealso{
\url{https://github.com/misasa/casteml}
}

