% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.R
\name{cbk.plot}
\alias{cbk.plot}
\title{Download CASTEML file and create geochemical diagram}
\usage{
cbk.plot(stone, category = "trace")
}
\arguments{
\item{stone}{Unique indentification number of stones in Medusa.}

\item{category}{Category filter that is passed to `casteml convert'}
}
\value{
Dataframe used to plot the diagram
}
\description{
Download CASTEML file and create geochemical diagram
}
\examples{
cbk.plot("20130318074936-009-271",category="trace")
cbk.plot("20081202172326.hkitagawa",category="trace")
cbk.plot("20081202172326.hkitagawa",category="lead")
cbk.plot("20130528105235-594267",category="oxygen")
}
\seealso{
\code{\link{cbk.plot.trace}}
}

