% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.ree.R
\name{cbk.plot.ree}
\alias{cbk.plot.ree}
\title{Read CASTEML file and create REE diagram}
\usage{
cbk.plot.ree(pmlfile, tableunit = "ug/g", property = "atomicnumber",
  reference = "Wasson.1988")
}
\arguments{
\item{pmlfile}{File path to CASTEML file}

\item{tableunit}{Unit to toss to cbk.read.casteml()}

\item{property}{Property to align x-axis}

\item{reference}{Reference of element abundance}
}
\value{
Dataframe used to plot the diagram
}
\description{
Read CASTEML file and create REE diagram.  This
  function does not save the created diagram.  You should prepare a
  canvas in advance.
}
\examples{
pmlfile <- cbk.download.casteml("20081202172326.hkitagawa")
cbk.plot.ree(pmlfile)
}
\seealso{
\url{https://github.com/misasa/casteml}
}

