% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.trace.R
\name{cbk.plot.trace}
\alias{cbk.plot.trace}
\title{Read dataframe and create trace diagram}
\usage{
cbk.plot.trace(pmlfile, tableunit = "ug/g", property = "atomicnumber",
  reference = "Wasson.1988")
}
\arguments{
\item{dataframe}{file path to dataframe}
}
\value{
file path to created file
}
\description{
Read dataframe and create trace diagram.  The
  created file is stored in current directory.  This function
  returns path to the file.
}
\examples{
pmlfile <- cbk.download.casteml("20081202172326.hkitagawa")
rplotfile <- sprintf("\%s.\%s",sub("\\\\.[[:alnum:]]+$","",pmlfile),"pdf")
pdf(rplotfile)
cbk.plot.trace(pmlfile)
dev.off()
cbk.exec(rplotfile)
}
\seealso{
\code{casteml download}, \url{https://github.com/misasa/casteml}, \code{\link{cbk.download.casteml}}
}

