% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.read.dataframe.R
\name{cbk.read.dataframe}
\alias{cbk.read.dataframe}
\title{Read dataframe created from casteml with column "unit"}
\usage{
cbk.read.dataframe(dataframefile, tableunit)
}
\arguments{
\item{dataframefile}{A flipped csvfile}

\item{tableunit}{Preferred unit that will be passed to `cbk.convector'}
}
\value{
A dataframe with unit organized
}
\description{
Read dataframe created from casteml with column "unit"
}
\details{
This is designed to read CASTEML file by Rscript.
}
\seealso{
\code{\link{read.csv}}
}

