% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.read.dflame.R
\name{cbk.read.dflame}
\alias{cbk.read.dflame}
\title{Read csvfile created by casteml convert with column `unit'}
\usage{
cbk.read.dflame(dflame.csv, tableunit = "none", verbose = TRUE)
}
\arguments{
\item{dflame.csv}{A csvfile with columns of stone and rows of element
abundances, with 2nd column `unit'.}

\item{tableunit}{Output unit that will be resolved by
\link{cbk.convector} (default: "none").}

\item{verbose}{Output debug info (default: TRUE).}
}
\value{
A pmlame with unit organized.
}
\description{
Read csvfile with column `unit' created by
  \code{casteml convert --format dataframe}.  This is low-level
  function and users are not encourage to call this but
  cbk.read.casteml().
}
\details{
This internally calls
  \code{read.csv(dflame.csv,row.names=1,header=T,stringsAsFactors=F)},
  take out column of `unit' and normalized by the `unit' column.
}
\examples{
pmlfile    <- cbk.download.casteml("20081202172326.hkitagawa")
dflame.csv <- cbk.convert.casteml(pmlfile,category="trace")
pmlame     <- cbk.read.dflame(dflame.csv,"ppm")

pmlame     <- cbk.read.dflame(cbk.path("20081202172326.hkitagawa_trace.dflame"),"ppm")
pmlame     <- cbk.read.dflame(cbk.path("ref1-dflame0.csv"),"ppm")
pmlame     <- cbk.read.dflame(cbk.path("periodic-dflame0.csv"))
}
\seealso{
\code{\link{cbk.download.casteml}}, \code{casteml
  convert}, \url{https://github.com/misasa/casteml}, and
  \code{\link{cbk.convector}}.
}

