% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.read.pmlame.R
\name{cbk.read.pmlame}
\alias{cbk.read.pmlame}
\title{Get and return a pmlame}
\usage{
cbk.read.pmlame(pmlfile_or_stone, opts = NULL, use_gem = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exists locally or
stone-ID.}

\item{opts}{List of further options for plot.}

\item{use_gem}{Download a CASTEML file into a temporary directory
(default: FALSE).}

\item{verbose}{Output debug info (default: TRUE).}
}
\value{
A dataframe with unit organized.
}
\description{
Get and return a pmlame.  This function can use two
  pathways to get a pmlame.  The one downloads a json data from
  medusa, while the other downloads a CASTEML file using
  rubygem. Finally, they are converted to pmlames with same
  structure.
}
\examples{
pmlfile <- cbk.path("20081202172326.hkitagawa.pml")
pmlame  <- cbk.read.pmlame(pmlfile,tableunit="ppm",category="trace")
stone   <- "20081202172326.hkitagawa"
pmlame  <- cbk.read.pmlame(stone,tableunit="ppm")
}
\seealso{
\code{\link{cbk.convert.casteml}},
  \code{\link{cbk.read.dflame}}, and
  \url{https://github.com/misasa/casteml}
}
