% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.read.R
\name{cbk.read}
\alias{cbk.read}
\title{Return a dataframe stored in Medusa}
\usage{
cbk.read(stone, tableunit = "none", category = "trace")
}
\arguments{
\item{stone}{Unique indentification number of stones in
Medusa.  Really, those will pass to `casteml download'.}

\item{tableunit}{Output unit that will be resolved by
`cbk.convector' (default="none")}

\item{category}{category to pass to `casteml convert'}
}
\value{
A dataframe with unit organized
}
\description{
Return a dataframe stored in Medusa.  This function
  downloads CASTEML file by external command `casteml' then read
  it as data.frame.
}
\details{
This function downloads CASTEML file by `casteml
  download', convert it to csvfile by `casteml convert', and read
  the file by `cbk.read.dataframe()'.
}
\examples{
tbl0 <- cbk.read("20081202172326.hkitagawa","ppm","trace")

}
\seealso{
\code{casteml download}, \code{casteml convert},
  \url{https://github.com/misasa/casteml},
  \code{\link{cbk.casteml.download}}, \code{\link{cbk.casteml.convert}},
  \code{\link{cbk.read.datafram}}
}

