% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.read.tblame.R
\name{cbk.read.tblame}
\alias{cbk.read.tblame}
\title{Read csvfile created by casteml convert with column `unit'}
\usage{
cbk.read.tblame(tblame, tableunit = "none", verbose = TRUE)
}
\arguments{
\item{tblame}{A csvfile with columns of chem and rows of stone,
with 2nd row `unit'}

\item{tableunit}{Output unit that will be resolved by
\link{cbk.convector} (default: "none").}

\item{verbose}{Output debug info (default: TRUE).}
}
\value{
A pmlame with unit organized.
}
\description{
Read csvfile with row `unit', that is also referred as
  `tblame'.
}
\details{
This internally calls \code{\link{read.csv}}, take out row
  of `unit' and normalized by the `unit' row.
}
\seealso{
\code{\link{cbk.read.dflame}},
  \url{https://github.com/misasa/casteml}, and
  \code{\link{cbk.convector}}.
}

