% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.read.yaml.R
\name{cbk.read.yaml}
\alias{cbk.read.yaml}
\title{Read a pmlame stored in yamlfile.}
\usage{
cbk.read.yaml(yamlfile, verbose = TRUE)
}
\arguments{
\item{yamlfile}{A yamlfile with pmlame, that is with columns of
chem and rows of stone.}

\item{verbose}{Output debug info (default: TRUE).}
}
\value{
A pmlame with columns of chem and rows of stone.
}
\description{
Read a pmlame stored in yamlfile.
}
\seealso{
\code{\link{cbk.write.yaml}},
  \code{\link{cbk.read.casteml}}.
}

