% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.reducer.R
\name{cbk.reducer}
\alias{cbk.reducer}
\title{Convert major-element concentration as oxide to one as metal}
\usage{
cbk.reducer(tbl0, oxides = NULL)
}
\arguments{
\item{tbl0}{A dataframe with columns of stone and rows of
element abundances}

\item{oxide}{A list of oxide to be converted}
}
\value{
A dataframe with converted columns
}
\description{
Convert major-element concentration as oxide to one as metal
}
\examples{
pmlfile <- cbk.download.casteml("20081202172326.hkitagawa")
tbl0    <- cbk.read.casteml(pmlfile,"ppm",category=NULL)
tbl1    <- cbk.reducer(tbl0)
}

