% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.ref.R
\name{cbk.ref}
\alias{cbk.ref}
\title{Return element abundances from file `ref.csv'}
\usage{
cbk.ref(analysis, property = NULL)
}
\arguments{
\item{analysis}{A name of reference that is either "Boynton.1989", "McDonough.1995", "Wasson.1988"}

\item{property}{A numeric vector of element property with label
that serves as sort key (default=NULL).  You can feed output from `cbk.periodic'}
}
\value{
A numeric vector of element abundances with label
}
\description{
Return element abundances from file `ref.csv'
}
\examples{
cbk.ref("Boynton.1989")
cbk.ref("Wasson.1988",cbk.periodic('atomicnumber'))
cbk.ref("McDonough.1995",cbk.periodic('volatility'))
}

