% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.texify.R
\name{cbk.texify}
\alias{cbk.texify}
\title{Convert CASTEML file to multple texfiles}
\usage{
cbk.texify(pmlfile_or_stone, outfile = "table-auto.tex", chem = NULL,
  ncol = 11, verbose = FALSE)
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exits locally or
stone-ID (or pmlame).}

\item{outfile}{File path to texfile.}

\item{chem}{List of chem such as c("Li","Si","Ca","Ca.1","Rb").}

\item{ncol}{Number of columns per a table.}

\item{verbose}{Output debug info.}
}
\value{
Vector of file path to texfiles.
}
\description{
Convert CASTEML file to multple texfiles.  This
  function splits chem and parse the datasets separatedly.
}
\seealso{
\code{\link{cbk.lame.texify1}}
}
