% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.vector.R
\name{cbk.vector}
\alias{cbk.vector}
\title{Convert dataframe with label to vector with label.  This is written
at initial stage of development.  May be not that useful.}
\usage{
cbk.vector(casdfm)
}
\arguments{
\item{casdfm}{A dataframe with label}
}
\value{
A vector with label
}
\description{
Convert dataframe with label to vector with label.  This is written
at initial stage of development.  May be not that useful.
}

