% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.write.casteml.R
\name{cbk.write.casteml}
\alias{cbk.write.casteml}
\title{Write a pmlame to CASTEML file}
\usage{
cbk.write.casteml(pmlame, outfile = NULL, digit = 6, verbose = TRUE)
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{outfile}{File path to casteml file}

\item{digit}{Integer indicating the number of significant digits to
be used by \code{signif}.}

\item{verbose}{Output debug info (default: TRUE).}
}
\value{
File path to a casteml file
}
\description{
Write a pmlame to CASTEML file
}
\examples{
pmlame0 <- data.frame(row.names=c('ref_cpx_klb1@1','ref_cpx_klb1@2','trc_meso_allende@10'),SiO2=c(520000,520000,600000))
pmlfile <- cbk.write.casteml(pmlame0,'deleteme.pml')
}
\seealso{
\code{\link{cbk.read.casteml}}
}
