% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.write.tblame.R
\name{cbk.write.tblame}
\alias{cbk.write.tblame}
\title{Write pmlame as TBLAME.csv}
\usage{
cbk.write.tblame(pmlame, outfile, verbose = TRUE)
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{outfile}{Path of TBLAME.csv, which is a csvfile with rows of
stone and columns of chem.}

\item{verbose}{Output debug info (default: FALSE).}
}
\value{
Path of TBLAME.csv
}
\description{
Write pmlame as TBLAME.csv
}
\seealso{
\code{\link{write.csv}}, \code{\link{cbk.read.tblame}},
  and \url{https://github.com/misasa/casteml}
}
