% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.write.yaml.R
\name{cbk.write.yaml}
\alias{cbk.write.yaml}
\title{Write a pmlame to yamlfile.}
\usage{
cbk.write.yaml(pmlame, outfile = NULL, digit = 6, verbose = TRUE)
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{outfile}{Path to yamlfile with a pmlame.}

\item{digit}{Integer indicating the number of significant digits to
be used by \code{signif}.}

\item{verbose}{Output debug info (default: TRUE).}
}
\value{
Path to yamlfile.
}
\description{
Write a pmlame to yamlfile.
}
\details{
This function uses \code{yaml::as.yaml} to save pmlame.
  Since \code{yaml::as.yaml} drops rownames, this function stores
  rownames using a column with label `rownames'.
}
\seealso{
\code{\link{cbk.read.yaml}} and
  \code{\link{cbk.write.casteml}}.
}
