% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chelyabinsk-package.R
\docType{package}
\name{chelyabinsk-package}
\alias{chelyabinsk}
\alias{chelyabinsk-package}
\title{Utility to plot geochemical data.}
\description{
This package allows you to plot geochemical data
}

