% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionml.convert.laicpqms.R
\name{ionml.convert.laicpqms}
\alias{ionml.convert.laicpqms}
\title{Convert laicpqms-type QTEGRA.CSV by Analyte G2 with iCAP-Q
  to ion-type TBLAME.CSV}
\usage{
ionml.convert.laicpqms(file, outfile = NULL, force = FALSE)
}
\arguments{
\item{file}{Name of laicpqms-type QTEGRA.CSV exported form Qtegra}

\item{outfile}{Name of ion-type TBLAME.csv file that will be
created}

\item{force}{Flag to force convert again}
}
\value{
Name of ion-type TBLAME.csv file that was saved
}
\description{
Convert laicpqms-type QTEGRA.CSV by Analyte G2 with
  iCAP-Q to ion-type TBLAME.CSV.  Filter metadata out and append
  column of cycle number.

  This does NOT produce IONML but does imcomplete IONML that is
  ion-type TBLAME.csv.
}
\examples{
file <- cbk.path("ref_cpx_klb1@1.qtg")
ionml.convert.laicpqms(file,outfile=tempfile(fileext=".ion"))
}
\seealso{
ionml-laicpqms.plx
}

