% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionml.read.session.R
\name{ionml.read.session}
\alias{ionml.read.session}
\title{Read a session stored as a series of icnml-type TBLAME.csv}
\usage{
ionml.read.session(tblame.csv, ref = "Si29")
}
\arguments{
\item{tblame.csv}{Paths to icnml-type TBLAME.csv}

\item{ref}{Reference ion such as `Si29'}
}
\value{
The icnml-type pmlame with average and standard error
}
\description{
Read a session stored as a series of icnml-type
  TBLAME.csv.  As of January 5, 2017, this function uses
  `ionml.read.laicpqms' to read the icnml-type TBLAME.csv.
}
\seealso{
\code{\link{ionml.read.laicpqms}}
}

