% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas.balance.R
\name{cas.balance}
\alias{cas.balance}
\title{Do mass balance calculation}
\usage{
cas.balance(datain, element, phase, mode, density)
}
\arguments{
\item{datain}{Element abundances with row "WR"}

\item{mode}{Vector of modal abundance of phases with label}

\item{density}{Vector of density of phases with label}

\item{elements}{Vector of string such for c("Ba","Nb","La"...)}

\item{phases}{Vector of name of phases such for c("ol", "pig")}
}
\value{
Element abundances with rows of remainder and series of
    quantity, and columns of density and mode
}
\description{
Do mass balance calculation
}
\examples{
> datain
          Ba       Nb       La       Ce       Pr        Sr       Nd       Sm
ol  0.044000 0.049000 0.000415 0.004941 0.000608  0.619333 0.000000 0.000000
pig 0.059048 0.183133 0.003487 0.016731 0.004169  0.546383 0.041143 0.043506
WR  4.510000 0.022500 0.013000 0.031000 0.004400 77.300000 0.029000 0.014000
          Zr       Hf       Eu       Gd       Dy       Li        Y       Er
ol  0.264633 0.024833 0.003022 0.004000 0.027036 1.635333 0.000000 0.023356
pig 0.370700 0.043394 0.006900 0.069043 0.180000 0.682633 1.370666 0.178516
WR  0.140000 0.005000 0.002800 0.032000 0.061000 2.050000 0.462000 0.057000
          Yb       Lu
ol  0.031766 0.010288
pig 0.200000 0.037148
WR  0.077000 0.014200
> element
 [1] "Ba" "Nb" "La" "Ce" "Pr" "Sr" "Nd" "Sm" "Zr" "Hf" "Eu" "Gd" "Dy" "Li" "Y"
[16] "Er" "Yb" "Lu"
> phase
[1] "ol"  "pig"
> mode
   ol   pig    WR
0.525 0.353 1.000
> density
 ol pig  WR
3.3 3.4 3.3
QQ <- cas.balance(datain,elements,phases,mode,density)
QQzCI <- cas.normalize(QQ,ref0)
}

