% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas.read.dataframe.R
\name{cas.read.dataframe}
\alias{cas.read.dataframe}
\title{Read dataframe created from casteml with column "unit"}
\usage{
cas.read.dataframe(dataframefile, tableunit)
}
\arguments{
\item{dataframefile}{A flipped csvfile}

\item{tableunit}{Preferred unit that will be passed to `cas.convector'}
}
\value{
A dataframe with unit organized
----------------
* unit definition (should be defined as function in the future)
----------------
----------------
* load dataset from csvfile with unit
----------------
}
\description{
Read dataframe created from casteml with column "unit"
}
\details{
This is designed to read CASTEML file by Rscript.
}
\seealso{
\code{\link{read.csv}}
}

