% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas.read.dataframe.R
\name{cas.read.dataframe}
\alias{cas.read.dataframe}
\title{Read dataframe created from casteml with column `unit'}
\usage{
cas.read.dataframe(dataframefile, tableunit = "none")
}
\arguments{
\item{dataframefile}{A flipped csvfile}

\item{tableunit}{Preferred unit, that is one of c("none","g/g","wt%","cg/g","%","permil","mg/g","ppm","ug/g","ppb","ng/g","pg/g")}
}
\value{
A dataframe
}
\description{
Read dataframe created from casteml with column `unit'
}
\examples{
\dontrun{
$ casteml download -R 20130528105235-594267 > 20130528105235-594267.pml
$ casteml convert -f dataframe -c trace 20130528105235-594267.pml > 20130528105235-594267.dataframe
R> tbl0 <- cas.read.dataframe("20130528105235-594267.dataframe","ppm")
}
----------------
* unit definition (should be defined as function in the future)
----------------
----------------
* load dataset from csvfile with unit
----------------
}

