% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas.ref.R
\name{cas.ref}
\alias{cas.ref}
\title{Return element abundances from file `ref.csv'}
\usage{
cas.ref(analysis, property = NULL)
}
\arguments{
\item{analysis}{A name of ANALYSIS such like "Boynton 1989"}

\item{property}{A numeric vector of element property with label
that serves as sort key (option)}
}
\value{
A numeric vector of element abundances with label
}
\description{
Return element abundances from file `ref.csv'
}
\examples{
> cas.ref("Boynton 1989")
    La     Ce     Pr     Nd     Sm     Eu     Gd     Tb     Dy     Ho     Er
0.3100 0.8080 0.1220 0.6000 0.1950 0.0735 0.2590 0.0474 0.3220 0.0718 0.2100
    Tm     Yb     Lu
0.0324 0.2090 0.0322
}

