% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas.ref.R
\name{cas.ref}
\alias{cas.ref}
\title{Return element abundances from file `ref.csv'}
\usage{
cas.ref(analysis, property = NULL)
}
\arguments{
\item{analysis}{A name of ANALYSIS such like "Boynton 1989"}

\item{property}{A numeric vector of element property with label
that serves as sort key (option)}
}
\value{
A numeric vector of element abundances with label
}
\description{
Return element abundances from file `ref.csv'
}
\examples{
cas.ref("Boynton 1989")
}

