% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.casteml.convert.R
\name{cbk.casteml.convert}
\alias{cbk.casteml.convert}
\title{Convert CASTEML file to CSV file}
\usage{
cbk.casteml.convert(pmlfile, category = "trace")
}
\arguments{
\item{pmlfile}{file path to CASTEML file}

\item{category}{category to pass to `casteml convert'}
}
\value{
file path to converted file in temporary directory
}
\description{
Convert CASTEML file to CSV file.  The converted file
  is stored in a temporary directory.  This function returns path
  to the file.
}
\examples{
pmlfile <- cbk.casteml.download("20081202172326.hkitagawa")
cbkfile <- cbk.casteml.convert(pmlfile,category="trace")

cbkfile <- cbk.casteml.convert(cbk.path("20081202172326.kitagawa.pml"),category="trace")
}
\seealso{
\code{casteml convert}, \url{https://github.com/misasa/casteml}, \code{\link{cbk.casteml.download}}
}

