% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.casteml.convert.R
\name{cbk.casteml.convert}
\alias{cbk.casteml.convert}
\title{Convert CASTEML files to temporary directory and return the file path}
\usage{
cbk.casteml.convert(pmlfile, category = "trace")
}
\arguments{
\item{pmlfile}{file path to CASTEML file}

\item{category}{category to pass to `casteml convert'}
}
\value{
file path to converted file in temporary directory
}
\description{
Convert CASTEML files to temporary directory and return the file path
}

