% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.casteml.download.R
\name{cbk.casteml.download}
\alias{cbk.casteml.download}
\title{Download analysis record in temporary directory and return the file path}
\usage{
cbk.casteml.download(stone)
}
\arguments{
\item{ID}{Unique indentification number of stones in Medusa}
}
\value{
file path to CASTEML file that was downloaded in temporary directory
}
\description{
Download analysis record in temporary directory and return the file path
}

