% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.category.R
\name{cbk.category}
\alias{cbk.category}
\title{Suggest categories to plot}
\usage{
cbk.category(pmlfile_or_stone = NULL)
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exits locally or
stone-ID (or pmlame) to survey categories.  Default is NULL and
return all categories regardless their plotability.}
}
\value{
Categories in preferred order
}
\description{
Suggest categories to plot
}
\examples{
pmlfile  <- cbk.path("20081202172326.hkitagawa.pml")
category <- cbk.category(pmlfile)
category <- cbk.category()
}

