% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.category.suggest.R
\name{cbk.category.suggest}
\alias{cbk.category.suggest}
\title{Suggest categories to plot}
\usage{
cbk.category.suggest(pmlfile_or_stone = NULL)
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exits locally or stone-ID (or pmlame)}
}
\value{
Categories in preferred order
}
\description{
Suggest categories to plot
}
\examples{
pmlfile  <- cbk.path("20081202172326.hkitagawa.pml")
category <- cbk.category.suggest(pmlfile)
category <- cbk.category.suggest()
}

