% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.convector.R
\name{cbk.convector}
\alias{cbk.convector}
\title{Return conversion cofficient for certain unit}
\usage{
cbk.convector(unit)
}
\arguments{
\item{unit}{Unit string that is one of none, g/g, wt\%, cg/g, \%,
permil, mg/g, ppm, ug/g, ppb, ng/g, pg/g}
}
\value{
Conversion cofficient.  When "permil", this returns 1000
  instead of 0.001.  Very likely you want to DIVIDE table by the
  factor
}
\description{
Return conversion cofficient for certain unit
}
\examples{
cbk.convector("none")
cbk.convector("permil")
cbk.convector("ppm")
}

