% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.convert.casteml.R
\name{cbk.convert.casteml}
\alias{cbk.convert.casteml}
\title{Convert CASTEML file to CSV file}
\usage{
cbk.convert.casteml(pmlfile, category = NULL, force = FALSE)
}
\arguments{
\item{pmlfile}{File path to CASTEML file}

\item{category}{Category filter that is passed to `casteml convert'}
}
\value{
File path to the converted file
}
\description{
Convert CASTEML file to CSV file.  This function
  returns path to the converted file.  The converted file is stored
  in a temporary directory.  Note with the same arguments, this
  function convert file only once per a R session.  This is
  low-level function and users are not encourage to call this but
  cbk.read.casteml().
}
\examples{
pmlfile <- cbk.path("20081202172326.hkitagawa.pml")
cbkfile <- cbk.convert.casteml(pmlfile)

pmlfile <- cbk.download.casteml("20081202172326.hkitagawa")
cbkfile <- cbk.convert.casteml(pmlfile,category="trace")
}
\seealso{
\code{casteml convert},
  \url{https://github.com/misasa/casteml}
}

