% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.convert.casteml.R
\name{cbk.convert.casteml}
\alias{cbk.convert.casteml}
\title{Convert CASTEML file to certain file}
\usage{
cbk.convert.casteml(pmlfile, category = NULL, force = FALSE,
  verbose = TRUE, format = "dataframe")
}
\arguments{
\item{pmlfile}{File path to CASTEML file.}

\item{category}{Category specifier that is passed to \code{casteml
convert}.}

\item{force}{Flag to force convert again.}

\item{verbose}{Output debug info (default: TRUE).}

\item{format}{Format specifier that is passed to \code{casteml
convert}.}
}
\value{
Path of the converted file.
}
\description{
Convert CASTEML file to certain file.  This function
  returns path to the converted file.  The converted file is stored
  in a temporary directory.  Note with the same arguments, this
  function convert file only once per a R session.  This is
  low-level function and users are not encourage to call this but
  cbk.read.casteml().
}
\examples{
pmlfile    <- cbk.path("20081202172326.hkitagawa.pml")
dflame.csv <- cbk.convert.casteml(pmlfile)

pmlfile    <- cbk.download.casteml("20081202172326.hkitagawa")
texfile    <- cbk.convert.casteml(pmlfile,format="tex")
}
\seealso{
\code{casteml convert} and
  \url{https://github.com/misasa/casteml}
}

