% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.exec.R
\name{cbk.exec}
\alias{cbk.exec}
\title{Open file even on Mac}
\usage{
cbk.exec(file)
}
\arguments{
\item{file}{A filename}
}
\value{
result by opening the file
}
\description{
Open file even on Mac
}
\examples{
cbk.exec(".")
}

