% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.exec.R
\name{cbk.exec}
\alias{cbk.exec}
\title{Open a file or URL using Windows/Macintosoh file associations}
\usage{
cbk.exec(file)
}
\arguments{
\item{file}{file or URL to be opened}
}
\value{
result by opening the file
}
\description{
Open the specified file or URL using the application
  specified in the Windows/Macintosh file associations.  On Windows
  and Macintosh, this calls \code{\link{shell.exec}} and
  \code{\link{system}}, respectively.
}
\details{
The path in \code{file} is interpreted relative to the
  current working directory.
}
\examples{
cbk.exec(".")
}
\seealso{
\code{\link{shell.exec}} and \code{\link{system}}
}

