% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.iso.R
\name{cbk.iso}
\alias{cbk.iso}
\title{Return properties of isotope}
\usage{
cbk.iso(isomeas = NULL, property = "mass")
}
\arguments{
\item{isomeas}{A name of isotope such as `Li7' and `Rb87'.}

\item{property}{A name of PROPERTY that is one of `proton',
`nucleon', `mass', or `ratio'}
}
\value{
A scalar of property or pmlame of isotope table
}
\description{
Return properties of isotope.  It is in a form of
  pmlame.
}
\examples{
cbk.iso()
cbk.iso('Li7')
cbk.iso('Li7','ratio')
}
\seealso{
\code{\link{cbk.periodic}}, isotope.m, isotope_constant.m
}

