% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.colMeans.R
\name{cbk.lame.colMeans}
\alias{cbk.lame.colMeans}
\title{Return mean value of each element}
\usage{
cbk.lame.colMeans(pmlame)
}
\arguments{
\item{pmlame}{A pmlame with rows of stone and columns of chem}
}
\value{
A pmlame with mean value of each column
}
\description{
Return mean value of each element
}
\examples{
pmlame  <- cbk.read.casteml(cbk.path("20130528105235-594267.pml"))
pmlame0 <- pmlame[,setdiff(colnames(pmlame), c("sample_id","image_id"))]
pmlame1 <- cbk.lame.colMeans(pmlame0)
}

