% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.colMeans.R
\name{cbk.lame.colMeans}
\alias{cbk.lame.colMeans}
\title{Return mean of each chem}
\usage{
cbk.lame.colMeans(pmlame, stonify_regexp = "@[[:digit:]]+$")
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{stonify_regexp}{A regexp string to be stripped off to stonefy
analyses. Matched string will be stripped off from acqlist to
have stonelist.}
}
\value{
A pmlame with mean value of each column
}
\description{
Return mean of each chem
}
\examples{
pmlfile <- cbk.path("20130528105235-594267.pml")
message(sprintf("The pmlfile is located at |\%s|.",pmlfile))
pmlame  <- cbk.read.casteml(pmlfile)
pmlame0 <- cbk.lame.regulate(pmlame,mean=TRUE,error=FALSE,extra=FALSE)
pmlame1 <- cbk.lame.colMeans(pmlame0)
}
