% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.colSds.R
\name{cbk.lame.colSds}
\alias{cbk.lame.colSds}
\title{Return standard deviation of each chem}
\usage{
cbk.lame.colSds(pmlame, groupfy_stone_regexp = "@[[:digit:]]+$")
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{groupfy_stone_regexp}{A regexp string to groupfy
stones. Actually it splits stone names to group name and sub-name
that would be ignored}
}
\value{
A pmlame with standard deviation value of each column
}
\description{
Return standard deviation of each chem
}
\examples{
pmlame  <- cbk.read.casteml(cbk.path("20130528105235-594267.pml"))
pmlame0 <- pmlame[,setdiff(colnames(pmlame), c("sample_id","image_id"))]
pmlame1 <- cbk.lame.colSds(pmlame0)
}
