% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.drop.dharma.R
\name{cbk.lame.drop.dharma}
\alias{cbk.lame.drop.dharma}
\title{Remove rows where all values are NA}
\usage{
cbk.lame.drop.dharma(pmlame, column = FALSE, verbose = FALSE)
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{column}{Flag to remove columns}

\item{verbose}{Output debug info (default: FALSE).}
}
\value{
A pmlame without removed rows
}
\description{
Remove rows where all values are NA
}
\examples{
pmlfile <- cbk.path("20081202172326.hkitagawa.pml")
pmlame  <- cbk.read.casteml(pmlfile,"ppm",category=NULL)
pmlame1 <- cbk.lame.drop.dharma(pmlame)
}
