% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.fetch.error.R
\name{cbk.lame.fetch.error}
\alias{cbk.lame.fetch.error}
\title{Extract error columns from pmlame}
\usage{
cbk.lame.fetch.error(pmlame, chem = NULL)
}
\arguments{
\item{pmlame}{A pmlame with both mean and error}

\item{chem}{Target element you want to extract}
}
\value{
An errorlame that is a pmlame of only error values with
  label of chem
}
\description{
Extract error columns from pmlame
}
\examples{
pmlame       <- cbk.read.casteml(cbk.path("20130528105235-594267.pml"))
pmlame_error <- cbk.lame.fetch.error(pmlame)

pmlame_error <- cbk.lame.fetch.error(pmlame,c("Li","Lu"))
}
