% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.fetch.error.R
\name{cbk.lame.fetch.error}
\alias{cbk.lame.fetch.error}
\title{Extract error columns from pmlame}
\usage{
cbk.lame.fetch.error(pmlame)
}
\arguments{
\item{pmlame}{A pmlame with values of both mean and error}
}
\value{
A pmlame of error values
}
\description{
Extract error columns from pmlame
}
\examples{
pmlame        <- cbk.read.casteml(cbk.path("20130528105235-594267.pml"))
pmlame0       <- pmlame[,colnames(pmlame) != c("sample_id","image_id")]
pmlame0_mean  <- cbk.lame.colMeans(pmlame0)
pmlame0_error <- cbk.lame.colSds(pmlame0)
pmlame1       <- cbk.lame.merge.error(pmlame0_mean,pmlame0_error)
pmlame1_error <- cbk.lame.fetch.error(pmlame1)
}

