% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.merge.error.R
\name{cbk.lame.merge.error}
\alias{cbk.lame.merge.error}
\title{Merge columns of mean and standard-error interleavely}
\usage{
cbk.lame.merge.error(pmlame, pmlame_error)
}
\arguments{
\item{pmlame}{A pmlame of mean values}

\item{pmlame_error}{A pmlame of error values}
}
\value{
A pmlame with merged columns
}
\description{
Merge columns of mean and standard-error interleavely
}
\examples{
pmlame            <- cbk.read.casteml(cbk.path("20130528105235-594267.pml"))
pmlame0           <- pmlame[,colnames(pmlame) != c("sample_id","image_id")]
pmlame1           <- cbk.lame.colMeans(pmlame0)
pmlame1_error     <- cbk.lame.colSds(pmlame0)
pmlame_with_error <- cbk.lame.merge.error(pmlame1,pmlame1_error)
}

