% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.merge.error.R
\name{cbk.lame.merge.error}
\alias{cbk.lame.merge.error}
\title{Merge errorlame to pmlame interleavely}
\usage{
cbk.lame.merge.error(meanlame, errorlame, verbose = FALSE)
}
\arguments{
\item{meanlame}{A pmlame of mean values with colname of chem.}

\item{errorlame}{An errorlame of error values with colname of chem.}

\item{verbose}{Output debug info (default: FALSE).}
}
\value{
A pmlame with values of both mean and error.
}
\description{
Merge errorlame to pmlame interleavely
}
\examples{
pmlame        <- cbk.read.casteml(cbk.path("20130528105235-594267.pml"))
pmlame0       <- pmlame[,!grepl("_error",colnames(pmlame))]
pmlame1       <- pmlame0[,setdiff(colnames(pmlame0), c("sample_id","image_id"))]
pmlame1_mean  <- cbk.lame.colMeans(pmlame1)
pmlame1_error <- cbk.lame.colSds(pmlame1)
pmlame2       <- cbk.lame.merge.error(pmlame1_mean,pmlame1_error)
}
\seealso{
\code{cbk.lame.error}
}
