% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.merge.error.R
\name{cbk.lame.merge.error}
\alias{cbk.lame.merge.error}
\title{Merge columns of mean and error interleavely}
\usage{
cbk.lame.merge.error(pmlame_mean, pmlame_error)
}
\arguments{
\item{pmlame_mean}{A pmlame of mean values}

\item{pmlame_error}{A pmlame of error values}
}
\value{
A pmlame with values of both mean and error
}
\description{
Merge columns of mean and error interleavely
}
\examples{
pmlame        <- cbk.read.casteml(cbk.path("20130528105235-594267.pml"))
pmlame0       <- pmlame[,!grepl("_error",colnames(pmlame))]
pmlame1       <- pmlame0[,setdiff(colnames(pmlame0), c("sample_id","image_id"))]
pmlame1_mean  <- cbk.lame.colMeans(pmlame1)
pmlame1_error <- cbk.lame.colSds(pmlame1)
pmlame2       <- cbk.lame.merge.error(pmlame1_mean,pmlame1_error)
}

