% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.merge.R
\name{cbk.lame.merge}
\alias{cbk.lame.merge}
\title{Merge to two pmlames}
\usage{
cbk.lame.merge(pmlame1, pmlame2, verbose = FALSE)
}
\arguments{
\item{pmlame1}{A pmlame with row of stone and column of chem [g/g].}

\item{pmlame2}{Another pmlame.}

\item{verbose}{Output debug info (default: FALSE).}
}
\value{
A combined pmlame with row of stone and column of chem
}
\description{
Merge to two pmlames
}
\seealso{
\code{cbk.lame.merge.error}
}

