% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.regulate.R
\name{cbk.lame.regulate}
\alias{cbk.lame.regulate}
\title{Regulate pmlame to extract mean, error, or extra columns}
\usage{
cbk.lame.regulate(pmlame, mean = TRUE, error = TRUE, extra = FALSE)
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{mean}{Flag to extract mean columns (default: TRUE).}

\item{error}{Flag to extract error columns (default: TRUE).}

\item{extra}{Flag to extract extra columns (default: FALSE).}
}
\value{
A pmlame with columns of concern
}
\description{
Regulate pmlame to extract mean, error, or extra columns
}
\examples{
pmlame  <- cbk.read.casteml(cbk.path("20130528105235-594267.pml"))
pmlame1 <- cbk.lame.regulate(pmlame)
}
