% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.regulate.R
\name{cbk.lame.regulate}
\alias{cbk.lame.regulate}
\title{Regulate pmlame to remove non-chem components}
\usage{
cbk.lame.regulate(pmlame)
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}
}
\value{
A pmlame with only chem
}
\description{
Regulate pmlame to remove non-chem components
}
\examples{
pmlame0 <- cbk.read.casteml(cbk.path("20130528105235-594267.pml"))
pmlame  <- cbk.lame.regulate(pmlame0)
}

