% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.regulate.R
\name{cbk.lame.regulate}
\alias{cbk.lame.regulate}
\title{Regulate pmlame to extract mean, error, or extra columns}
\usage{
cbk.lame.regulate(pmlame, mean = TRUE, error = TRUE, extra = FALSE,
  chem = NULL)
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{mean}{Flag to extract mean columns (default: TRUE).}

\item{error}{Flag to extract error columns (default: TRUE).  Note
that dataframe return by option(mean=F,error=T,extra=F) is NOT
errorlame.  To get errorlame, use \code{cbk.lame.fetch.error}
instead.}

\item{extra}{Flag to extract extra columns (default: FALSE).}

\item{chem}{Select only certain chem for mean and error.  With chem
c('Li','Sr88'), columns of 'Li','Li_error','Sr88','Sr88_error'
are selected.}
}
\value{
A pmlame with columns of concern
}
\description{
Regulate pmlame to extract mean, error, or extra columns
}
\examples{
pmlfile <- cbk.path("20130528105235-594267.pml")
message(sprintf("The pmlfile is located at |\%s|.",pmlfile))
pmlame  <- cbk.read.casteml(pmlfile)
pmlame1 <- cbk.lame.regulate(pmlame)
}
\seealso{
\code{\link{cbk.lame.fetch.error}}
}
