% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.lame.texfy.R
\name{cbk.lame.texfy}
\alias{cbk.lame.texfy}
\title{Convert pmlame to texfile}
\usage{
cbk.lame.texfy(pmlame, chem, outfile = NULL)
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem [g/g].}

\item{chem}{List of chem such as c("Li","Si","Ca","Ca.1","Rb")}

\item{outfile}{File path to texfile}
}
\value{
File path to texfile
}
\description{
Convert pmlame to texfile
}
\examples{
pmlfile <- cbk.path("ref_phase_obj.pml")
pmlame0 <- cbk.read.casteml(pmlfile)
chem1   <- c("Li","B","Si","Rb","Sr","Y","Zr","Nb","Cs")
cbk.lame.texfy(pmlame0,chem1)
}

