% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.normalize.R
\name{cbk.normalize}
\alias{cbk.normalize}
\title{Return normalized element abundances}
\usage{
cbk.normalize(pmlame, ref, suffix_after_name_of_element = NULL)
}
\arguments{
\item{pmlame}{A dataframe of element abundances of sample}

\item{ref}{A dataframe of element abundances of reference}

\item{suffix_after_name_of_element}{String to recognize column of
errors.  Feed "_error" when necessary (default: NULL)}
}
\value{
A ref-normalized daraframe with only elements defined in
  ref
}
\description{
Return normalized element abundances.  Note that only
  elements that exist both in pmlame and ref are processed.  See
  also "Geochemical Modelling..." by Janousek et al. (2015)
}
\examples{
pmlame <- cbk.read.dflame(cbk.path("20081202172326.hkitagawa_trace.dataframe"),"ppm")
ref    <- cbk.ref("Boynton.1989","ppm")
cbk.normalize(pmlame,ref)
}
\seealso{
\code{\link{cbk.ref}}, \code{\link{cbk.periodic}}
}

