% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.path.R
\name{cbk.path}
\alias{cbk.path}
\title{Convert filename of local library to absolute path}
\usage{
cbk.path(name)
}
\arguments{
\item{name}{A filename}
}
\value{
A path to lib directory
}
\description{
Convert filename of local library to absolute path
}
\examples{
cbkfile <- cbk.path("periodic-table1.dataframe")
cbkfile <- cbk.path("ref1.dataframe")
pmlfile <- cbk.path("20081202172326.kitagawa.pml")
cbkfile <- cbk.path("20081202172326.kitagawa_trace.dataframe")
}

