% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.path.R
\name{cbk.path}
\alias{cbk.path}
\title{Convert filename of local library to absolute path}
\usage{
cbk.path(name)
}
\arguments{
\item{name}{A filename}
}
\value{
A path to lib directory
}
\description{
Convert filename of local library to absolute path
}
\examples{
cbk.path("periodic-table.csv")
cbk.path("ref.csv")
cbk.path("ref1.dataframe")
}

