% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.path.R
\name{cbk.path}
\alias{cbk.path}
\title{Return absolute path of files stored in local library}
\usage{
cbk.path(filename)
}
\arguments{
\item{filename}{A filename}
}
\value{
An absolute path of files stored in local library
}
\description{
Return absolute path of files stored in local library
}
\examples{
dflame.csv <- cbk.path("periodic-dflame0.csv")
dflame.csv <- cbk.path("ref1-dflame0.csv")
pmlfile    <- cbk.path("20081202172326.hkitagawa.pml")
dflame.csv <- cbk.path("20081202172326.hkitagawa_trace.dflame")
}

