% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.balance.R
\name{cbk.plot.balance}
\alias{cbk.plot.balance}
\title{Draw results of mass balance calculation}
\usage{
cbk.plot.balance(pmlame_or_file, modeinfo, reflame, verbose = TRUE)
}
\arguments{
\item{pmlame_or_file}{A pmlame or CASTEML file with element
abundance of phases without WR}

\item{modeinfo}{A CSV file with rows of mode and density and columns
of phases including WR}

\item{reflame}{A dataframe with element abundance of WR}

\item{verbose}{Output debug info (default: FALSE).}
}
\value{
A dataframe with relavie quantity of element.
}
\description{
Draw results of mass balance calculation.
}
\examples{
reflame <- cbk.ref("Wasson.1988",property=cbk.periodic("atomicnumber"))
MM      <- cbk.plot.balance(cbk.path("demo-trc0.pml"),cbk.path("demo-mode0.csv"),reflame)
}
