% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.balance.R
\name{cbk.plot.balance}
\alias{cbk.plot.balance}
\title{Balance quantity of elements in whole rock and phases}
\usage{
cbk.plot.balance(pmlame_or_file, reflame_or_file, modeinfo, verbose = FALSE)
}
\arguments{
\item{pmlame_or_file}{A pmlame or CASTEML or TBLAME.csv with
element abundance of phases without measured WR}

\item{reflame_or_file}{A pmlame or CASTEML or TBLAME.csv with
element abundance of measured WR}

\item{modeinfo}{A pmlame or TBLAME.csv with mode and density of
phases including WR}

\item{verbose}{Output debug info (default: FALSE).}
}
\value{
A pmlame with element quantities and mode of phases
}
\description{
Balance quantity of elements in whole rock and phases.
  This calculates quantity of elements in each phases from mode and
  element abundances of each phase.  Then from calculated
  whole-rock (WR) quantity of element by summation and measured WR
  quantity of element, this estimates surplus or deficit of
  elements.
}
\examples{
reflame <- cbk.ref("Anders.1989",property=cbk.periodic("atomicnumber"))
MM      <- cbk.plot.balance(cbk.path("demo-trc0.pml"),reflame,cbk.path("demo-mode0.csv"))
}
