% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.lead.R
\name{cbk.plot.lead}
\alias{cbk.plot.lead}
\title{Read CASTEML file and create lead diagram}
\usage{
cbk.plot.lead(pmlfile_or_stone, opts = NULL)
}
\arguments{
\item{pmlfile_or_stone}{A dataframe of element abundances (or pmlfile or stone-ID).}

\item{opts}{List of further options for plot.  See
\code{\link{cbk.plot}}.}
}
\value{
A pmlmae used to plot the diagram.
}
\description{
Read CASTEML dataframe and create lead diagram.  This
  function does not save the created diagram.  You should prepare a
  canvas in advance.
}
\examples{
pmlfile <- cbk.path("20081202172326.hkitagawa.pml")
cbk.plot.lead(pmlfile)
}
\seealso{
\url{https://github.com/misasa/casteml}
}

