% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.lithium.R
\name{cbk.plot.lithium}
\alias{cbk.plot.lithium}
\title{Read CASTEML file and create lithium diagram}
\usage{
cbk.plot.lithium(pmlfile_or_stone, opts = NULL)
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exits locally or stone-ID (or pmlame)}

\item{opts}{List of further options for plot}
}
\value{
Dataframe used to plot the diagram
}
\description{
Read CASTEML dataframe and create lithium diagram.
  This function does not save the created diagram.  You should
  prepare a canvas in advance.
}
\examples{
pmlfile <- cbk.path("20130528105235-594267.pml")
cbk.plot.lithium(pmlfile)
}
\seealso{
\url{https://github.com/misasa/casteml}
}

