% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.message.R
\name{cbk.plot.message}
\alias{cbk.plot.message}
\title{Display a diagram with message}
\usage{
cbk.plot.message(pmlfile_or_stone, msg)
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exits locally or stone-ID (or pmlame)}

\item{msg}{Message to be displayed on diagram}
}
\value{
Dataframe used to plot the diagram
}
\description{
Display a diagram with message
}
\examples{
pmlame <- cbk.read.casteml(cbk.path("20081202172326.hkitagawa.pml"),"ppm")
cbk.plot.message(pmlame,"Hello, world!")
}

