% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.oxygen.R
\name{cbk.plot.oxygen}
\alias{cbk.plot.oxygen}
\title{Read CASTEML file and create oxygen diagram}
\usage{
cbk.plot.oxygen(pmlfile_or_stone, opts = NULL)
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exits locally or stone-ID (or pmlame).}

\item{opts}{List of further options for plot.  See
\code{\link{cbk.plot}}.}
}
\value{
A pmlame used to plot the diagram.
}
\description{
Read CASTEML dataframe and create oxygen diagram.
  This function does not save the created diagram.  You should
  prepare a canvas in advance.
}
\examples{
pmlfile <- cbk.path("20130528105235-594267.pml")
cbk.plot.oxygen(pmlfile)
}
\seealso{
\url{https://github.com/misasa/casteml}
}
