% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.R
\name{cbk.plot}
\alias{cbk.plot}
\title{Plot Analyses with descendants of a stone}
\usage{
cbk.plot(pmlfile_or_stone, opts = NULL, category = NULL)
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exits locally or
stone-ID (or pmlame).}

\item{opts}{List of further options for plot.  Two options such as
`list(legendp=TRUE, Recursivep=FALSE)' are available.}

\item{category}{Category filter that is passed to `casteml
convert'.}
}
\value{
A pmlame used to plot the diagram.
}
\description{
Plot Analyses with descendants of a stone.  This
  function downloads CASTEML file associated with a stone and
  creates geochemical diagram depending on category specified.
}
\examples{
cbk.plot(cbk.path("20081202172326.hkitagawa.pml"))
cbk.plot(cbk.path("20081202172326.hkitagawa.pml"),category="trace")
cbk.plot(cbk.path("20081202172326.hkitagawa.pml"),category="lead")
cbk.plot(cbk.path("20130528105235-594267.pml"),category="trace")
cbk.plot(cbk.path("20130528105235-594267.pml"),category="lithium")
cbk.plot(cbk.path("20130528105235-594267.pml"),category="oxygen")
cbk.plot(cbk.path("20081202172326.hkitagawa.pml"),opts=list(legendp=FALSE))
}
\seealso{
\code{\link{cbk.plot.trace}}
}

