% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.R
\name{cbk.plot}
\alias{cbk.plot}
\title{Plot Analyses with descendants of a stone}
\usage{
cbk.plot(stone, category = "default")
}
\arguments{
\item{stone}{Unique indentification number of stone in Medusa.
This will be passed to cbk.download.casteml() with "--recursive"
option.}

\item{category}{Category filter that is passed to `casteml convert'}
}
\value{
Dataframe used to plot the diagram
}
\description{
Plot Analyses with descendants of a stone.  This
  function downloads CASTEML file associated with a stone and
  creates geochemical diagram depending on category specified.
}
\examples{
cbk.plot("20130318074936-009-271",category="trace")
cbk.plot("20081202172326.hkitagawa",category="trace")
cbk.plot("20130528105235-594267",category="lithium")
cbk.plot("20130528105235-594267",category="oxygen")
cbk.plot("20081202172326.hkitagawa",category="lead")
}
\seealso{
\code{\link{cbk.plot.trace}}
}

