% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.R
\name{cbk.plot}
\alias{cbk.plot}
\title{Plot Analyses with descendants of a stone}
\usage{
cbk.plot(pmlfile_or_stone, opts = NULL, category = NULL, ...)
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exits locally or
stone-ID (or pmlame).}

\item{opts}{List of further options for plot.  Two options such as
`list(legendp=TRUE, Recursivep=FALSE)' are available.}

\item{category}{Category specifier that is passed to \code{casteml
convert}.}
}
\value{
A pmlame used to plot the diagram.
}
\description{
Plot Analyses with descendants of a stone.  This
  function downloads CASTEML file associated with a stone and
  creates geochemical diagram depending on category specified.
}
\examples{
cbk.plot(c("20141106120733-593074","20141106120433-380274","20141106120509-983968"))

pmlfile <- cbk.path("20081202172326.hkitagawa.pml")
message(sprintf("The pmlfile is located at |\%s|.",pmlfile))
cbk.plot(pmlfile)
cbk.plot(pmlfile,category="trace")
cbk.plot(pmlfile,category="lead")
cbk.plot(pmlfile,opts=list(legendp=FALSE))

pmlfile <- cbk.path("20130528105235-594267.pml")
message(sprintf("The pmlfile is located at |\%s|.",pmlfile))
cbk.plot(pmlfile,category="trace")
cbk.plot(pmlfile,category="lithium")
cbk.plot(pmlfile,category="oxygen")
}
\seealso{
\code{\link{cbk.plot.trace}}
}
