% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.R
\name{cbk.plot}
\alias{cbk.plot}
\title{Plot Analyses with descendants of a stone}
\usage{
cbk.plot(pmlfile_or_stone, category = "default", opts = NULL)
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exits locally or stone-ID (or pmlame)}

\item{category}{Category filter that is passed to `casteml convert'}
}
\value{
Dataframe used to plot the diagram
}
\description{
Plot Analyses with descendants of a stone.  This
  function downloads CASTEML file associated with a stone and
  creates geochemical diagram depending on category specified.
}
\examples{
cbk.plot(cbk.path("20081202172326.hkitagawa.pml"))
cbk.plot(cbk.path("20081202172326.hkitagawa.pml"),category="trace")
cbk.plot(cbk.path("20081202172326.hkitagawa.pml"),category="lead")
cbk.plot(cbk.path("20130528105235-594267.pml"),category="trace")
cbk.plot(cbk.path("20130528105235-594267.pml"),category="lithium")
cbk.plot(cbk.path("20130528105235-594267.pml"),category="oxygen")
}
\seealso{
\code{\link{cbk.plot.trace}}
}

