% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.ree.R
\name{cbk.plot.ree}
\alias{cbk.plot.ree}
\title{Read CASTEML file and create REE diagram}
\usage{
cbk.plot.ree(pmlfile_or_stone, opts = NULL, tableunit = "none",
  reference = "Wasson.1988", verbose = FALSE, pch = NULL, col = NULL,
  ...)
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exits locally or
stone-ID (or pmlame).}

\item{opts}{List of further options for plot.  See \link{cbk.plot}.}

\item{tableunit}{Unit to toss to \link{cbk.read.casteml}.}

\item{reference}{Reference of element abundance.}

\item{verbose}{Output debug info (default: FALSE).}

\item{pch}{Array of symbol (default: NULL)}

\item{col}{Array of color (default: NULL)}
}
\value{
@return A pmlame used to plot the diagram.
}
\description{
Read CASTEML dataframe and create REE diagram.  This
  function tris to blow errorbar.  If you do not see errorbar,
  check pmlame that you feed.
}
\examples{
pmlfile <- cbk.path("20081202172326.hkitagawa.pml")
cbk.plot.ree(pmlfile)

pmlame  <- cbk.read.casteml(cbk.path("20160921173604-511857.pml"))
cbk.plot.ree(pmlame)
}
\seealso{
\url{https://github.com/misasa/casteml}
}
