% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.spider.R
\name{cbk.plot.spider}
\alias{cbk.plot.spider}
\title{Read CASTEML file and create spider diagram}
\usage{
cbk.plot.spider(pmlfile_or_stone, opts = NULL, tableunit = "none",
  property = "atomicnumber", reference = "Wasson.1988")
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exits locally or
stone-ID (or pmlame).}

\item{opts}{List of further options for plot.  See
\code{\link{cbk.plot}}.}

\item{tableunit}{Unit to toss to \link{cbk.read.casteml}.}

\item{property}{Property to align x-axis to pass to \code{\link{cbk.periodic}}.}

\item{reference}{Reference of element abundance}
}
\value{
A pmlame used to plot the diagram.
}
\description{
Read CASTEML dataframe and create spider diagram.
  This function does not save the created diagram.  You should
  prepare a canvas in advance.
}
\examples{
pmlfile <- cbk.path("20081202172326.hkitagawa.pml")
pmlame  <- cbk.read.casteml(pmlfile,tableunit="none",category=NULL)
cbk.plot.spider(pmlame)

pmlfile <- cbk.path("20160921173604-511857.pml")
pmlame  <- cbk.read.casteml(pmlfile,tableunit="none",category=NULL)
cbk.plot.spider(pmlame)
}
\seealso{
\url{https://github.com/misasa/casteml}
}

