% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.spots.R
\name{cbk.plot.spots}
\alias{cbk.plot.spots}
\title{Read CASTEML file and create spots diagram}
\usage{
cbk.plot.spots(pmlfile_or_surface, opts = NULL, imagefile = NULL)
}
\arguments{
\item{pmlfile_or_surface}{A CASTEML file that exits locally or
stone-ID (or pmlame).}

\item{opts}{List of further options for plot.  See
\code{\link{cbk.plot}}.}

\item{imagefile}{Image shown on background of analyzed spots.}
}
\value{
A pmlame used to plot the diagram.
}
\description{
Read CASTEML dataframe and create spots diagram.
  This function does not save the created diagram.  You should
  prepare a canvas in advance.
}
\examples{
pmlfile <- cbk.download.casteml("20160819165624-372633")
cbk.plot.spots(pmlfile)
}
\seealso{
\url{https://github.com/misasa/casteml}
}

