% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.trace.R
\name{cbk.plot.trace}
\alias{cbk.plot.trace}
\title{Read CASTEML file and create trace diagram}
\usage{
cbk.plot.trace(pmlfile, tableunit = "ug/g", property = "atomicnumber",
  reference = "Wasson.1988")
}
\arguments{
\item{pmlfile}{File path to CASTEML file}
}
\value{
trace diagram
}
\description{
Read CASTEML file and create trace diagram.  This
  function does not save the created diagram.  You should prepare a
  canvas in advance.
}
\examples{
pmlfile <- cbk.download.casteml("20081202172326.hkitagawa")
cbk.plot.trace(pmlfile)
}
\seealso{
\url{https://github.com/misasa/casteml}
}

