% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.trace.R
\name{cbk.plot.trace}
\alias{cbk.plot.trace}
\title{Read CASTEML file and create trace diagram}
\usage{
cbk.plot.trace(pmlfile_or_stone, opts = NULL, tableunit = "none")
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exits locally or
stone-ID (or pmlame).}

\item{opts}{List of further options for plot.  See
\code{\link{cbk.plot}}.}

\item{tableunit}{Unit to toss to \code{\link{cbk.read.casteml}}.}
}
\value{
A pmlame used to plot the diagram.
}
\description{
Read CASTEML dataframe and create trace diagram.  This
  function does not save the created diagram.  You should prepare a
  canvas in advance.
}
\examples{
pmlfile <- cbk.path("20081202172326.hkitagawa.pml")
pmlame  <- cbk.read.casteml(pmlfile,tableunit="ug/g",category=NULL)
cbk.plot.trace(pmlame,tableunit="ug/g")
}
\seealso{
\url{https://github.com/misasa/casteml}
}

