% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.plot.trace.R
\name{cbk.plot.trace}
\alias{cbk.plot.trace}
\title{Read CASTEML file and create trace diagram}
\usage{
cbk.plot.trace(pmlame, tableunit = "ug/g", property = "atomicnumber",
  reference = "Wasson.1988")
}
\arguments{
\item{pmlame}{A dataframe of element abundances (or pmlfile or stone-ID)}

\item{tableunit}{Unit to toss to cbk.read.casteml()}

\item{property}{Property to align x-axis}

\item{reference}{Reference of element abundance}
}
\value{
Dataframe used to plot the diagram
}
\description{
Read CASTEML dataframe and create trace diagram.  This
  function does not save the created diagram.  You should prepare a
  canvas in advance.
}
\examples{
pmlfile <- cbk.path("20081202172326.hkitagawa.pml")
pmlame  <- cbk.read.casteml(pmlfile,tableunit="ug/g",category=NULL)
cbk.plot.trace(pmlame)
}
\seealso{
\url{https://github.com/misasa/casteml}
}

