% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.read.casteml.R
\name{cbk.read.casteml}
\alias{cbk.read.casteml}
\title{Read CASTEML file and return a dataframe}
\usage{
cbk.read.casteml(pmlfile_or_stone, tableunit = "none", category = NULL,
  download = FALSE)
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exits locally or
stone-ID (when download=TRUE)}

\item{tableunit}{Output unit that will be resolved by
cbk.convector() (default="none")}

\item{category}{category to pass to `cbk.convert.casteml'}

\item{download}{flag to directory download from Medusa
(default=FALSE)}
}
\value{
A dataframe with unit organized
}
\description{
Read CASTEML dataframe and return a dataframe.
}
\details{
This function converts a CASTEML file to a csvfile by
  `cbk.convert.casteml()' and read it by `cbk.read.dataframe()'.
  When `download' option was specified, this directly reads stones
  from Medusa.
}
\examples{
pmlfile <- cbk.path("20081202172326.hkitagawa.pml")
pmlame  <- cbk.read.casteml(pmlfile,tableunit="ppm",category="trace")

pmlame  <- cbk.read.casteml("20081202172326.hkitagawa",tableunit="ppm",category="trace",download=TRUE)
}
\seealso{
\code{\link{cbk.convert.casteml}},
  \code{\link{cbk.read.dataframe}},
  \url{https://github.com/misasa/casteml}
}

