% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.read.casteml.R
\name{cbk.read.casteml}
\alias{cbk.read.casteml}
\title{Read CASTEML file and return a dataframe}
\usage{
cbk.read.casteml(pmlfile, tableunit = "none", category = "trace")
}
\arguments{
\item{pmlfile}{CASTEML file download by `casteml download'}

\item{tableunit}{Output unit that will be resolved by
`cbk.convector()' (default="none")}

\item{category}{category to pass to `cbk.casteml.convert'}
}
\value{
A dataframe with unit organized
}
\description{
Read CASTEML file and return a dataframe.
}
\details{
This function converts CASTEML file to csvfile by
  `cbk.casteml.convert()' and read it by `cbk.read.dataframe()'.
}
\examples{
pmlfile <- cbk.path("20081202172326.kitagawa.pml")
tbl0 <- cbk.read.casteml(pmlfile,"ppm","trace")

}
\seealso{
\code{\link{cbk.casteml.convert}},
  \code{\link{cbk.read.dataframe}},
  \url{https://github.com/misasa/casteml}
}

