% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.read.casteml.R
\name{cbk.read.casteml}
\alias{cbk.read.casteml}
\title{Read CASTEML file and return a pmlame}
\usage{
cbk.read.casteml(pmlfile_or_stone, opts = NULL, tableunit = "none",
  category = NULL, force = FALSE, verbose = TRUE)
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exits locally or
stone-ID.}

\item{opts}{List of further options for plot.}

\item{tableunit}{Output unit that will be resolved by
cbk.convector() (default: "none").}

\item{category}{Category specifier that is passed to \code{casteml
convert}.}

\item{force}{Force read pmlfile with duplicated acquisitions
(default: FALSE).}

\item{verbose}{Output debug info (default: TRUE).}
}
\value{
A dataframe with unit organized.
}
\description{
Read CASTEML file and return a pmlame.
}
\details{
This function converts a CASTEML file to a csvfile by
  `cbk.convert.casteml()' and read it by `cbk.read.dflame()'.
}
\examples{
pmlfile <- cbk.path("20081202172326.hkitagawa.pml")
pmlame  <- cbk.read.casteml(pmlfile,tableunit="ppm",category="trace")
stone   <- "20081202172326.hkitagawa"
pmlame  <- cbk.read.casteml(stone,tableunit="ppm",category="trace")
}
\seealso{
\code{\link{cbk.convert.casteml}},
  \code{\link{cbk.read.dflame}}, and
  \url{https://github.com/misasa/casteml}
}
