% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.read.dataframe.R
\name{cbk.read.dataframe}
\alias{cbk.read.dataframe}
\title{Read csvfile created by casteml convert with column `unit'}
\usage{
cbk.read.dataframe(csvfile, unit)
}
\arguments{
\item{csvfile}{A csvfile with columns of stone and rows of element
abundances, with 2nd column `unit'}

\item{unit}{Preferred unit that will be passed to
`cbk.convector'}
}
\value{
A dataframe with unit organized
}
\description{
Read csvfile created by \code{casteml convert --format dataframe} with
  column `unit'.
}
\details{
This internally calls
  \code{read.csv(csvfile,row.names=1,header=T,stringsAsFactors=F)}.
}
\examples{
cbk.read.dataframe(cbk.path("20130528105235-594267_trace.dataframe"),"ppm")
}
\seealso{
\url{http://dream.misasa.okayama-u.ac.jp}, \code{\link{cbk.convector}}
}

