% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.read.dataframe.R
\name{cbk.read.dataframe}
\alias{cbk.read.dataframe}
\title{Read csvfile created by casteml convert with column `unit'}
\usage{
cbk.read.dataframe(csvfile, tableunit = "none")
}
\arguments{
\item{csvfile}{A csvfile with columns of stone and rows of element
abundances, with 2nd column `unit'}

\item{tableunit}{Output unit that will be resolved by
cbk.convector() (default="none")}
}
\value{
A dataframe with unit organized
}
\description{
Read csvfile with column `unit' created by
  \code{casteml convert --format dataframe} .  This is low-level
  function and users are not encourage to call this but
  cbk.read.casteml().
}
\details{
This internally calls
  \code{read.csv(csvfile,row.names=1,header=T,stringsAsFactors=F)},
  take out column of `unit' and normalized by the `unit' column.
}
\examples{
pmlfile <- cbk.download.casteml("20081202172326.hkitagawa")
cbkfile <- cbk.convert.casteml(pmlfile,category="trace")
tbl0    <- cbk.read.dataframe(cbkfile,"ppm")

tbl0    <- cbk.read.dataframe(cbk.path("20081202172326.hkitagawa_trace.dataframe"),"ppm")
tbl0    <- cbk.read.dataframe(cbk.path("ref1.dataframe"),"ppm")
tbl0    <- cbk.read.dataframe(cbk.path("periodic-table1.dataframe"))
}
\seealso{
\code{\link{cbk.download.casteml}}, \code{casteml
  convert}, \url{https://github.com/misasa/casteml},
  \code{\link{cbk.convector}},
}

