% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.read.medusa.R
\name{cbk.read.medusa}
\alias{cbk.read.medusa}
\title{Return analyses records stored in Medusa in a form of a
  dataframe}
\usage{
cbk.read.medusa(stone, tableunit = "none", category = NULL)
}
\arguments{
\item{stone}{Unique indentification number of stones in Medusa.
Really, those will pass to cbk.casteml.download().}

\item{tableunit}{Output unit that will be resolved by
cbk.convector() (default="none")}

\item{category}{category to pass to `casteml convert'}
}
\value{
A dataframe with unit organized
}
\description{
Return analyses records stored in Medusa in a form of
  a dataframe.  This function downloads CASTEML file by external
  command `casteml' then reads it as data.frame.
}
\details{
This function downloads a CASTEML file by
  `cbk.casteml.download()', converts it to csvfile by
  `cbk.casteml.convert()', and reads the file by
  `cbk.read.dataframe()'.
}
\examples{
tbl0 <- cbk.read.medusa("20081202172326.hkitagawa","ppm","trace")

}
\seealso{
\url{https://github.com/misasa/casteml},
  \code{\link{cbk.casteml.download}},
  \code{\link{cbk.casteml.convert}},
  \code{\link{cbk.read.dataframe}}
}

