% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.read.pmlame.R
\name{cbk.read.pmlame}
\alias{cbk.read.pmlame}
\title{Return a pmlame that is stored in Medusa}
\usage{
cbk.read.pmlame(pmlfile_or_stone, opts = NULL, casteml = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{pmlfile_or_stone}{A CASTEML file that exists locally or
stone-ID.}

\item{opts}{List of further options for plot.}

\item{casteml}{Download a CASTEML file using external program `gem package --
casteml' (default: FALSE).}

\item{verbose}{Output debug info (default: TRUE).}
}
\value{
A dataframe with unit organized.
}
\description{
Return a pmlame that is stored in Medusa.  There can
  be two approach to get the pmlame.  (1) The first approach is to
  downloads a CASTEML file using external program `gem package --
  casteml'.  (2) The other approach is to download a json data
  directoly from Medusa.

  The essential functions involved in this function as of May 29
  (2018) are shown below.

  cbk.read.pmlame <returns pmlame>
  --> cbk.read.casteml <returns pmlame>
      --> cbk.download.casteml <returns pmlfile>
  --> cbk.download.pmlame <returns pmlame>
      --> medusaRClient.read.pmlame <returns pmlame>
}
\examples{
pmlfile <- cbk.path("20081202172326.hkitagawa.pml")
message(sprintf("The pmlfile is located at |\%s|.",pmlfile))
pmlame  <- cbk.read.pmlame(pmlfile,tableunit="ppm",category="trace")
stone   <- "20081202172326.hkitagawa"
pmlame  <- cbk.read.pmlame(stone,tableunit="ppm")
}
\seealso{
\code{\link{cbk.convert.casteml}},
  \code{\link{cbk.read.dflame}}, and
  \url{https://github.com/misasa/casteml}
}
