% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.read.tblame.R
\name{cbk.read.tblame}
\alias{cbk.read.tblame}
\title{Read csvfile created by casteml convert with column `unit'}
\usage{
cbk.read.tblame(tblame, tableunit = "none")
}
\arguments{
\item{tblame}{A csvfile with columns of element abundances and
rows of stones, with 2nd row `unit'}

\item{tableunit}{Output unit that will be resolved by
cbk.convector() (default="none")}
}
\value{
A dataframe with unit organized
}
\description{
Read csvfile with row `unit', that is also referred as
  `tblame'.
}
\details{
This internally calls
  \code{read.csv(tblame,row.names=1,header=T,stringsAsFactors=F)},
  take out row of `unit' and normalized by the `unit' row.
}
\seealso{
\code{\link{cbk.read.dflame}},
  \url{https://github.com/misasa/casteml},
  \code{\link{cbk.convector}},
}

