% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.vector.R
\name{cbk.vector}
\alias{cbk.vector}
\title{Convert dataframe with label to vector with label}
\usage{
cbk.vector(casdfm)
}
\arguments{
\item{casdfm}{A dataframe with label}
}
\value{
A vector with label
}
\details{
This is written at initial stage of development.  May be
  not that useful.
}
\examples{
cbk.vector(data.frame(row.names=c("density"), ol=c(3.300), pig=c(3.400), WR=c(3.3)))
}

