% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.vector.R
\name{cbk.vector}
\alias{cbk.vector}
\title{Convert dataframe with label to vector with label}
\usage{
cbk.vector(df)
}
\arguments{
\item{df}{A dataframe with label}
}
\value{
A vector with label
}
\description{
Convert dataframe with label to vector with label.  It
  seems that this conversion is necessary for calcuration.
}
\details{
This is written at initial stage of development.  May be
  not that useful.
}
\examples{
cbk.vector(data.frame(row.names=c("density"), ol=c(3.300), pig=c(3.400), WR=c(3.3)))
}

