% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbk.write.yaml.R
\name{cbk.write.yaml}
\alias{cbk.write.yaml}
\title{Write a pmlame to yamlfile.}
\usage{
cbk.write.yaml(pmlame, yamlfile = NULL, verbose = TRUE)
}
\arguments{
\item{pmlame}{A pmlame with rows of stone and columns of chem.}

\item{yamlfile}{A yamlfile with a pmlame.}

\item{verbose}{Output debug info (default: TRUE).}
}
\value{
A pmlame with columns of chem and rows of stone.
}
\description{
Write a pmlame to yamlfile.
}
\details{
This function uses \link{yaml::as.yaml} to save pmlame.
  Since \link{yaml::as.yaml} drops rownames, this function stores
  rownames using a column with label `rownames'.
}
\seealso{
\code{\link{cbk.read.yaml}} and
  \code{\link{cbk.write.casteml}}.
}

