% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chelyabinsk-package.R
\docType{package}
\name{chelyabinsk-package}
\alias{chelyabinsk}
\alias{chelyabinsk-package}
\title{Express element distribution using comprehensive geochemical datasets}
\description{
Chelyabinsk allows you to handle comprehensive geochemical dataset.
As a first step, you plot REE and SPIDER diagram using appropriate
normalization.  Then, using both in-situ and whole-rock datasets,
you obtain element quantitative-distribution among phases with
mass-balance calculation.
}

