% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionml.convert.iontblame.R
\name{ionml.convert.iontblame}
\alias{ionml.convert.iontblame}
\title{Convert ion-type TBLAME.CSV to IONML}
\usage{
ionml.convert.iontblame(iontblame, outfile = NULL, force = FALSE)
}
\arguments{
\item{iontblame}{Name of ion-type TBLAME.CSV}

\item{outfile}{Name of IONML that will be created}

\item{force}{Flag to force convert again}
}
\value{
Name of IONML that was created
}
\description{
Convert ion-type TBLAME.CSV to IONML.

  This really create complate IONML.  This program assumes
  extensions of ion-type TBLAME.CSV and IONML to be `.ion' and
  `.xml', respectively.
}
\examples{
iontblame <- cbk.path("ref_cpx_klb1@1.ion")
ionml.convert.iontblame(iontblame,outfile=tempfile(fileext=".xml"))
}
\seealso{
ionml.convert.laicpqms
}

