% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionml.convert.laicpqms.R
\name{ionml.convert.laicpqms}
\alias{ionml.convert.laicpqms}
\title{Convert laicpqms-type QTEGRA.csv by Analyte G2 with iCAP-Q
  to IONCSV}
\usage{
ionml.convert.laicpqms(acqfile, outfile = NULL, force = FALSE)
}
\arguments{
\item{acqfile}{Name of laicpqms-type QTEGRA.csv exported form Qtegra}

\item{outfile}{Name of IONCSV that will be created}

\item{force}{Flag to force convert again}
}
\value{
Name of IONCSV that was created
}
\description{
Convert laicpqms-type QTEGRA.csv by Analyte G2 with
  iCAP-Q to IONCSV.  Filter metadata out and append column of cycle
  number.

  This program assumes extensions of laicpqms-type QTEGRA.csv and
  IONCSV to be `.csv' and `.ion', respectively.

  The IONCSV consists of columns of time and ion intensities.  The
  first column of each line should be number of `cycle'.  Colname
  of the IONCSV should be `time' and name of element followed by
  atomic weight (`Si29' instead of `29Si').
}
\examples{
acqfile <- cbk.path("ref_cpx_klb1@1.csv")
ionml.convert.laicpqms(acqfile,outfile=tempfile(fileext=".ion"))
}
\seealso{
ionml-laicpqms.plx
}
