% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionml.rate.delivery.R
\name{ionml.rate.delivery}
\alias{ionml.rate.delivery}
\title{Rate delivery of ion from a pit by laser abbration}
\usage{
ionml.rate.delivery(pmlame, intlame, radius = 1.25e-05, density = 3,
  depth = 3e-05, t = 40)
}
\arguments{
\item{pmlame}{A pmlame with row of stone and column of chem}

\item{intlame}{A pseudo-pmlame of ion intensity with row of stone
and column of isomeas [cps]}

\item{radius}{radius of a pit [m] (default: 12.5e-6)}

\item{density}{density of solid target [g/cc] (default: 3)}

\item{depth}{depth of a pit [m] (default: 30e-6)}

\item{t}{period of irradiation [s] (default: 40)}
}
\value{
Yield from solid target to detector
}
\description{
Rate delivery of ion from a pit by laser abbration.
  Although dimension of a pit is on meter [m], internaly, they are
  are converted into centi meter.
}

