% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionml.read.laicpqms.R
\name{ionml.read.laicpqms}
\alias{ionml.read.laicpqms}
\title{Read ion-type TBLAME.CSV originated from Analyte G2 with
  iCAP-Q}
\usage{
ionml.read.laicpqms(pmlame_or_file, t0 = 5, t1 = 20, t2 = 25, t3 = 60,
  ref = "Si29")
}
\arguments{
\item{pmlame_or_file}{ion-type pmlame or ion-type TBLAME.CSV with
extension `.ion'}

\item{t0}{time when baseline starts (default: 5 s)}

\item{t1}{time when baseline ends (default: 20 s)}

\item{t2}{time when ion starts (default: 25 s)}

\item{t3}{time when ion ends (default: 60 s)}

\item{ref}{reference ion such as "Si29"}
}
\value{
The ion-type pmlame of ion/ref online with rows of
  statistical information
}
\description{
Read ion-type TBLAME.CSV originated from Analyte G2
  with iCAP-Q.  The original QTEGRA.CSV (CSV file exported from
  Qtegra) should be processed in advance to have ion-type
  TBLNAME.CSV with extension `.ion'.

  The ion-type TBLNAME.CSV consists of columns of time and ion
  intensities.  The first column of each line should be number of
  `cycle'.  Colname of the ion-type TBLNAME.CSV should be `time'
  and name of element followed by atomic weight (`Si29' instead of
  `29Si').
}
\examples{
file <- cbk.path("ref_cpx_klb1@1.ion")
pmlfile0 <- ionml.read.laicpqms(file)
}
\seealso{
\code{\link{ionml.convert.laicpqms}}
}

