% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionml.read.laicpqms.R
\name{ionml.read.laicpqms}
\alias{ionml.read.laicpqms}
\title{Read IONML originated from Analyte G2 with iCAP-Q}
\usage{
ionml.read.laicpqms(pmlame_or_file, t0 = 5, t1 = 20, t2 = 25,
  t3 = 60, ref = "Si29", verbose = FALSE, ionml = TRUE)
}
\arguments{
\item{pmlame_or_file}{ion-type pmlame or IONML or QTEGRACSV.}

\item{t0}{When baseline starts (default: 5 s).}

\item{t1}{When baseline ends (default: 20 s).}

\item{t2}{When ion starts (default: 25 s).}

\item{t3}{When ion ends (default: 60 s).}

\item{ref}{reference ion such as `Si29'.}

\item{verbose}{Output debug info (default: FALSE).}

\item{ionml}{Read IONML file instead of IONCSV (default: TRUE).}
}
\value{
The ion-type pmlame of ion-to-ref ratio online with rows of
  statistical information.
}
\description{
Read IONML originated from Analyte G2 with iCAP-Q.
  When IONML is not found, this will create IONML from QTEGRACSV
  (CSV file exported from Qtegra with identical basename with
  IONML) via IONCSV stored in temporal dirctory.

  The IONCSV consists of columns of time and ion intensities.  The
  first column of each line should be number of `cycle'.  Colname
  of the IONCSV should be `time' and name of element followed by
  atomic weight (`Si29' instead of `29Si').

  This is a fork from `Batch_calc_separated_pdf_2.3.R'.  On
  2018-07-27, `summarise_each' was replaced by `summarise_all'.
}
\details{
Signal between `t2' and `t3' is regarded as main signal
  from a sample (online).  Signal between `t0' and `t1' is regarded
  as background.  Mean of latter is calculated as BASELINE.  Then
  the BASELINE is subtracted from the main signal.  The main signal
  is normalized by `ref'.  This function returns the BASELINE
  subtracted and reference normalized `ionic-ratio' with
  statistical information at the bottom of the table.  Detection
  limit is defined by 3 times standard error of BASELINE measument.
}
\examples{
ionmlfile <- cbk.path("ref_cpx_klb1@1.xml")
message(sprintf("The ionmlfile is located at |\%s|.",ionmlfile))
pmlfile0  <- ionml.read.laicpqms(ionmlfile)

file     <- cbk.path("ref_cpx_klb1@1.ion")
message(sprintf("The file is located at |\%s|.",file))
pmlfile0 <- ionml.read.laicpqms(file,ionml=FALSE)
}
\seealso{
\code{\link{ionml.convert.qtegracsv2ioncsv}}
  \code{\link{ionml.convert.ioncsv}}
}
