% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionml.read.laicpqms.R
\name{ionml.read.laicpqms}
\alias{ionml.read.laicpqms}
\title{Read ion-type TBLAME.csv originated from Analyte G2 with
  iCAP-Q}
\usage{
ionml.read.laicpqms(pmlame_or_file, t0 = 5, t1 = 20, t2 = 25, t3 = 60,
  ref = "Si29", verbose = FALSE)
}
\arguments{
\item{pmlame_or_file}{ion-type pmlame or ion-type TBLAME.csv with
extension `.ion'.}

\item{t0}{When baseline starts (default: 5 s).}

\item{t1}{When baseline ends (default: 20 s).}

\item{t2}{When ion starts (default: 25 s).}

\item{t3}{When ion ends (default: 60 s).}

\item{ref}{reference ion such as `Si29'.}

\item{verbose}{Output debug info (default: FALSE).}
}
\value{
The ion-type pmlame of ion-to-ref ratio online with rows of
  statistical information.
}
\description{
Read ion-type TBLAME.csv originated from Analyte G2
  with iCAP-Q.  The original QTEGRA.csv (CSV file exported from
  Qtegra) should be processed in advance to have ion-type
  TBLNAME.csv with extension `.ion'.

  The ion-type TBLNAME.csv consists of columns of time and ion
  intensities.  The first column of each line should be number of
  `cycle'.  Colname of the ion-type TBLNAME.csv should be `time'
  and name of element followed by atomic weight (`Si29' instead of
  `29Si').
}
\details{
Signal between `t2' and `t3' is regarded as main signal
  from a sample (online).  Signal between `t0' and `t1' is regarded
  as background.  Mean of latter is calculated as BASELINE.  Then
  the BASELINE is subtracted from the main signal.  The main signal
  is normalized by `ref'.  This function returns the BASELINE
  subtracted and reference normalized `ionic-ratio' with
  statistical information at the bottom of the table.  Detection
  limit is defined by 3 times standard error of BASELINE measument.
}
\examples{
file <- cbk.path("ref_cpx_klb1@1.ion")
pmlfile0 <- ionml.read.laicpqms(file)
}
\seealso{
\code{\link{ionml.convert.laicpqms}}
}
