% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionml.read.session.R
\name{ionml.read.session}
\alias{ionml.read.session}
\title{Read a session stored as a series of ion-type TBLAME.csv}
\usage{
ionml.read.session(tblame.csv, t0 = 5, t1 = 20, t2 = 25, t3 = 60,
  ref = "Si29", error = "sterr", intensityp = FALSE, verbose = FALSE)
}
\arguments{
\item{tblame.csv}{Paths to ion-type TBLAME.csv.}

\item{t0}{Time when baseline starts (default: 5 s, scalar or
vector)}

\item{t1}{Time when baseline ends (default: 20 s, scalar or
vector).}

\item{t2}{Time when ion starts (default: 25 s, scalar or vector).}

\item{t3}{Time when ion ends (default: 60 s, scalar or vector).}

\item{ref}{Reference ion such as `Si29'.}

\item{error}{Use `sterr', `DL', or `mix' as error.  When `mix',
error is delection limit `DL' instead of standard error `sderr'
when DL is larger than `sderr'.}

\item{intensityp}{Have mean and sd of intensity instead of mean and
sderr of ionic ratio}

\item{verbose}{Output debug info (default: FALSE).}
}
\value{
Pseudo-pmlame with average and standard error of ionic
  ratio
}
\description{
Read a session stored as a series of ion-type
  TBLAME.csv.  As of January 5, 2017, this function uses
  `ionml.read.laicpqms' to read the ion-type TBLAME.csv.
}
\examples{
files <- c(cbk.path("ref_cpx_klb1@1.ion"),cbk.path("ref_cpx_klb1@2.ion"),cbk.path("ref_cpx_klb1@3.ion"))
pmlame0 <- ionml.read.session(files)
}
\seealso{
\code{\link{ionml.read.laicpqms}}
}
