% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionml.read.session.R
\name{ionml.read.session}
\alias{ionml.read.session}
\title{Read a session stored as a series of ion-type TBLAME.csv}
\usage{
ionml.read.session(tblame.csv, ref = "Si29")
}
\arguments{
\item{tblame.csv}{paths to ion-type TBLAME.csv}

\item{ref}{reference ion such as `Si29'}
}
\value{
ion-type pmlame with average and standard error
}
\description{
Read a session stored as a series of ion-type
  TBLAME.csv.  As of January 5, 2017, this function uses
  `ionml.read.laicpqms' to read the ion-type TBLAME.csv.
}
\examples{
files <- c(cbk.path("ref_cpx_klb1@1.ion"),cbk.path("ref_cpx_klb1@2.ion"),cbk.path("ref_cpx_klb1@3.ion"))
pmlame0 <- ionml.read.session(files)
}
\seealso{
\code{\link{ionml.read.laicpqms}}
}

